refactor(calendar): CalendarStore counter → $state, fix delete ordering
Replaces the Writable<number> counter-bump pattern with a $state
version field, eliminating the $effect+subscribe bridge in
Calendar.svelte that existed because $derived.by() does not track
Svelte-store auto-subscriptions. The class moves to its own
calendarStore.svelte.ts file; store.ts keeps the pure fileMapKey and
computeFileMap helpers so tests continue to import them directly.

Also fixes a latent event-ordering bug (present in the old code too):
on delete, NoteCache's handler runs before CalendarStore's, so the
entry is already out of the cache by the time isPeriodic(file.path)
is checked — causing bump() to short-circuit and the calendar dot to
stay visible until the user navigated away and back. Delete events
now use bumpUnconditionally; getPeriodicNote self-heals any stale
entries in the subsequent re-derivation.

Closes #102.
