+++ THEORY.md
@@ -16,9 +16,7 @@ The core entities are few. A **granularity** is one of four time scales: day, we
 
 **Settings are a plain object, not reactive state.** The v2 refactor made this explicit: `plugin.settings` is a `Settings` object, mutated directly and persisted via `saveData()`. There is no Svelte store wrapping it. When settings change, the plugin fires a custom workspace event (`periodic-notes:settings-updated`) on Obsidian's event bus, and interested parties — the cache, the calendar store — subscribe to that event. This is a deliberate choice to keep Svelte confined to the calendar sidebar and nowhere else.
 
-**The calendar is a view into the cache, not a separate data structure.** The Svelte calendar components never query the vault directly. They go through `CalendarStore`, which wraps the plugin's cache lookups in a Svelte-compatible notification mechanism. `CalendarStore.store` is a `Writable<number>` — a counter whose value doesn't matter, only its changes. When a vault event might have changed the periodic note landscape, the counter bumps, and the `$effect` in `Calendar.svelte` re-computes a `FileMap`: a `Map<string, TFile | null>` that pre-resolves every day, week, month, and year cell that the current calendar view needs. This is the FileMap pattern — a single derivation replaces what would otherwise be 42+ individual store subscriptions (one per day cell, plus week/month/year). Child components like `Day.svelte` do a synchronous `$derived` lookup from this map using `fileMapKey()`.
-
-**The reactivity bridge is the most subtle piece of the architecture.** Svelte 5 runes (`$derived.by()`) do not automatically track Svelte 4 store subscriptions. The `Calendar.svelte` component works around this with a specific pattern: an `$effect` that manually calls `.subscribe()` on the store and returns the unsubscribe function (lines 44–54). This is not a bug or a transitional hack — it's the documented interop pattern for using Svelte 4 stores inside Svelte 5 runes mode. Issue #102 tracks a future migration to pure runes, at which point the `CalendarStore.store` counter would become a `$state` signal and the bridge would disappear.
+**The calendar is a view into the cache, not a separate data structure.** The Svelte calendar components never query the vault directly. They go through `CalendarStore` (in `calendarStore.svelte.ts`), which wraps the plugin's cache lookups and exposes a `$state` version counter — a number whose value doesn't matter, only its changes. When a vault event might have changed the periodic note landscape, the counter bumps, and the `$effect` in `Calendar.svelte` re-computes a `FileMap`: a `Map<string, TFile | null>` that pre-resolves every day, week, month, and year cell that the current calendar view needs. This is the FileMap pattern — a single derivation replaces what would otherwise be 42+ individual subscriptions (one per day cell, plus week/month/year). Child components like `Day.svelte` do a synchronous `$derived` lookup from this map using `fileMapKey()`.
 
 **Template rendering is layered by granularity.** `applyTemplate()` processes tokens in a fixed order: universal tokens (`{{date}}`, `{{title}}`, `{{time}}`), then granularity-specific tokens. Day notes get `{{yesterday}}`, `{{tomorrow}}`, and dynamic `{{date+1d:format}}`. Week notes get `{{monday:format}}` through `{{sunday:format}}`. Month and year notes get `{{month+1m:format}}` and `{{year-1y:format}}`. The key design choice: tokens for one granularity are left unreplaced in another granularity's template. `{{yesterday}}` in a month note template is preserved as literal text, not expanded. This is correct — it means templates are not accidentally granularity-polymorphic.
 
@@ -64,4 +62,4 @@ The core entities are few. A **granularity** is one of four time scales: day, we
 
 **The `isGranularityEnabled()` method in CalendarStore defaults to `true` for day granularity when settings are missing** (`?? granularity === "day"`, line 62). This fallback doesn't exist anywhere else in the codebase — in the cache, commands, and settings, a disabled granularity is simply skipped. The calendar apparently wants to always show day cells even if the day granularity is technically disabled in settings. This might be intentional (a calendar without day cells is useless) or might be a leftover guard from before settings loading was reliable.
 
-**The v2 refactor was a single-author, big-bang rewrite.** The commit history from `a1e4a5a` through `0ba5d35` shows a carefully sequenced but monolithic refactor: types first, then constants, then format, then template, cache, commands, settings, main, calendar updates, old module deletion, test rewrites. The design spec was committed, reviewed, amended, then deleted after completion. The spec listed issues it would close (#85, #93, #100, #102, #110). Of those, #93 (test re-implementation drift) and #102 (Svelte store migration) remain open — the refactor addressed them partially but not fully. The remaining open issues (#115, #116, #117, #118) are all optimization and testing work that the refactor deferred. The codebase is in a stable post-refactor state, but the testing boundary problem — the inability to import `cache.ts` and `template.ts` in tests — is the most significant piece of unfinished structural work.
+**The v2 refactor was a single-author, big-bang rewrite.** The commit history from `a1e4a5a` through `0ba5d35` shows a carefully sequenced but monolithic refactor: types first, then constants, then format, then template, cache, commands, settings, main, calendar updates, old module deletion, test rewrites. The design spec was committed, reviewed, amended, then deleted after completion. The spec listed issues it would close (#85, #93, #100, #102, #110). The remaining post-refactor work (#93, #102, #115, #116, #117, #118) was cleared in follow-up PRs: pure-logic extraction (`cacheIndex.ts`, `cacheSearch.ts`, `templateRender.ts`) made the test boundary problem tractable, and the calendar migrated fully to Svelte 5 runes. The codebase is in a stable state.
