+++ main.js
@@ -1,27 +1,27 @@
-let e=require(`obsidian`);function t(e,t){return`${e}:${t.clone().startOf(e).toISOString()}`}function n(e,t,n){let r=0,i=e.length-1;for(;r<=i;){let a=r+i>>>1,o=e[a];if(o===t)return e[a+(n===`forwards`?1:-1)]??null;o<t?r=a+1:i=a-1}return null}var r=[`day`,`week`,`month`,`year`],i=class{byPath=new Map;byKey=new Map;sortedByGranularity=new Map;dirtyGranularities=new Set(r);set(e){let n=t(e.granularity,e.date),r=this.byPath.get(e.filePath);if(r){let e=t(r.granularity,r.date);e!==n&&(this.byKey.delete(e),this.dirtyGranularities.add(r.granularity))}let i=this.byKey.get(n);i&&i.filePath!==e.filePath&&this.byPath.delete(i.filePath);let a=!this.byKey.has(n);this.byPath.set(e.filePath,e),this.byKey.set(n,e),a&&this.dirtyGranularities.add(e.granularity)}remove(e){let n=this.byPath.get(e);n&&(this.byKey.delete(t(n.granularity,n.date)),this.byPath.delete(e),this.dirtyGranularities.add(n.granularity))}clear(){this.byPath.clear(),this.byKey.clear(),this.sortedByGranularity.clear(),this.dirtyGranularities=new Set(r)}get(e){return e?this.byPath.get(e)??null:null}getByKey(e,n){return this.byKey.get(t(e,n))??null}has(e,t){let n=this.byPath.get(e);return n?t?t===n.granularity:!0:!1}findAdjacent(e,r){let i=this.get(e);if(!i)return null;let a=n(this.getSortedKeys(i.granularity),t(i.granularity,i.date),r);return a?this.byKey.get(a)??null:null}getSortedKeys(e){if(!this.dirtyGranularities.has(e)){let t=this.sortedByGranularity.get(e);if(t)return t}let t=`${e}:`,n=[];for(let e of this.byKey.keys())e.startsWith(t)&&n.push(e);return n.sort(),this.sortedByGranularity.set(e,n),this.dirtyGranularities.delete(e),n}},a={day:`YYYY-MM-DD`,week:`gggg-[W]ww`,month:`YYYY-MM`,year:`YYYY`},o={enabled:!1,format:``,folder:``,templatePath:void 0},s={granularities:{day:{...o},week:{...o},month:{...o},year:{...o}}},c=[`sunday`,`monday`,`tuesday`,`wednesday`,`thursday`,`friday`,`saturday`],l=`calendar`,u=Symbol(`displayedMonth`);function d(e,t){return e.granularities[t].format||a[t]}function f(e,t){let n=e.granularities[t].format;if(!n)return[a[t]];let r=/[^/]*$/.exec(n);return r?[n,r[0]]:[n]}function p(e,t){return e.granularities[t]}function m(e){return r.filter(t=>e.granularities[t].enabled)}function h(e){return e.replace(/\[[^\]]*\]/g,``).replace(/\\./g,``)}function g(e){return e.indexOf(`/`)===-1?e:e.split(`/`).pop()??``}function _(e){return!/[?<>\\:*|"]/g.test(e)&&!/[\x00-\x1f\x80-\x9f]/g.test(e)&&!/^\.+$/.test(e)&&!/^(con|prn|aux|nul|com[0-9]|lpt[0-9])(\..*)?$/i.test(e)}function v(e,t){if(!e)return``;if(!_(e))return`Format contains illegal characters`;if(t===`day`){let t=window.moment().format(e);if(!window.moment(t,e,!0).isValid())return`Failed to parse format`}return``}function y(e){let t=g(e).replace(/\[[^\]]*\]/g,``);return![`M`,`D`].every(e=>t.includes(e))||!(t.includes(`Y`)||t.includes(`y`))}function b(e,t){let n=window.moment().format(e);return window.moment(n,e,!0).isValid()?h(e).includes(`/`)&&t===`day`&&y(e)?`fragile-basename`:`valid`:`loose-parsing`}function ee(...e){let t=[];for(let n=0,r=e.length;n<r;n++)t=t.concat(e[n].split(`/`));let n=[];for(let e=0,r=t.length;e<r;e++){let r=t[e];!r||r===`.`||n.push(r)}return t[0]===``&&n.unshift(``),n.join(`/`)}var x=/{{\s*(date|time)\s*(([-+]\d+)([ymwdhs]))?\s*(:.+?)?}}/gi,te=/{{\s*(month)\s*(([-+]\d+)([ymwdhs]))?\s*(:.+?)?}}/gi,ne=/{{\s*(year)\s*(([-+]\d+)([ymwdhs]))?\s*(:.+?)?}}/gi,re=RegExp(`{{\\s*(${c.join(`|`)})\\s*:(.*?)}}`,`gi`);function ie(){let{moment:e}=window,t=e.localeData().firstDayOfWeek(),n=[...c];for(;t;){let e=n.shift();e&&n.push(e),t--}return n}function ae(e){let t=ie().indexOf(e.toLowerCase());return Math.max(0,t)}function oe(e,t,n,r,i){let a=window.moment();return e.replace(n,(e,n,o,s,c,l)=>{let u=t.clone();return i&&u.startOf(i),u.set({hour:a.get(`hour`),minute:a.get(`minute`),second:a.get(`second`)}),o&&u.add(parseInt(s,10),c),l?u.format(l.substring(1).trim()):u.format(r)})}function se(e,t,n,r,i){let a=i.replace(/{{\s*date\s*}}/gi,e).replace(/{{\s*time\s*}}/gi,window.moment().format(`HH:mm`)).replace(/{{\s*title\s*}}/gi,e);return t===`day`&&(a=a.replace(/{{\s*yesterday\s*}}/gi,n.clone().subtract(1,`day`).format(r)).replace(/{{\s*tomorrow\s*}}/gi,n.clone().add(1,`d`).format(r)),a=oe(a,n,x,r)),t===`week`&&(a=a.replace(re,(e,t,r)=>{let i=ae(t);return n.weekday(i).format(r.trim())})),t===`month`&&(a=oe(a,n,te,r,`month`)),t===`year`&&(a=oe(a,n,ne,r,`year`)),a}async function ce(t,n,r){if(!n||n===`/`)return``;let{metadataCache:i,vault:a}=t,o=(0,e.normalizePath)(n);try{let e=i.getFirstLinkpathDest(o,``);return e?a.cachedRead(e):``}catch(t){return console.error(`[Periodic Notes] Failed to read the ${r} note template '${o}'`,t),new e.Notice(`Failed to read the ${r} note template`),``}}async function le(e,t,n,r){let i=d(n,r.granularity),a=await ce(e,n.granularities[r.granularity].templatePath,r.granularity),o=se(t.basename,r.granularity,r.date,i,a);await e.vault.modify(t,o)}async function ue(t,n,r){let i=(0,e.normalizePath)(ee(r.folder??``,n.endsWith(`.md`)?n:`${n}.md`));return await de(t,i),i}async function de(e,t){let n=t.replace(/\\/g,`/`).split(`/`);n.pop();let r=``;for(let t of n)r=r?`${r}/${t}`:t,e.vault.getAbstractFileByPath(r)||await e.vault.createFolder(r)}function fe(e){let t=e.extension.length+1;return e.path.slice(0,-t)}function pe(e,t,n){if(b(t,n)===`fragile-basename`){let n=fe(e),r=(h(t).match(/\//g)?.length??0)+1;return n.split(`/`).slice(-r).join(`/`)}return e.basename}var me=class extends e.Component{index=new i;constructor(t,n){super(),this.app=t,this.plugin=n,this.app.workspace.onLayoutReady(()=>{console.info(`[Periodic Notes] initializing cache`),this.initialize(),this.registerEvent(this.app.vault.on(`create`,t=>{t instanceof e.TFile&&this.resolve(t,`create`)})),this.registerEvent(this.app.vault.on(`delete`,t=>{t instanceof e.TFile&&this.index.remove(t.path)})),this.registerEvent(this.app.vault.on(`rename`,this.onRename,this)),this.registerEvent(this.app.metadataCache.on(`changed`,this.onMetadataChanged,this)),this.registerEvent(this.app.workspace.on(`periodic-notes:settings-updated`,this.reset,this))})}reset(){console.info(`[Periodic Notes] resetting cache`),this.index.clear(),this.initialize()}initialize(){let t=this.plugin.settings,n=new Set,r=(t,i)=>{if(!n.has(t)){n.add(t);for(let n of t.children)n instanceof e.TFile?i(n):n instanceof e.TFolder&&r(n,i)}},i=m(t);for(let n of i){let i=t.granularities[n].folder||`/`,a=this.app.vault.getAbstractFileByPath(i);a instanceof e.TFolder&&r(a,t=>{if(t instanceof e.TFile){this.resolve(t,`initialize`);let e=this.app.metadataCache.getFileCache(t);e&&this.onMetadataChanged(t,``,e)}})}}onMetadataChanged(t,n,r){let i=this.plugin.settings,a=m(i);if(a.length!==0)for(let n of a){let a=i.granularities[n].folder||`/`;if(!t.path.startsWith(a===`/`?``:`${a}/`))continue;let o=(0,e.parseFrontMatterEntry)(r.frontmatter,n);if(!o)continue;let s=d(i,n);if(typeof o==`string`){let e=window.moment(o,s,!0);e.isValid()&&this.index.set({filePath:t.path,date:e,granularity:n,match:`frontmatter`});return}}}onRename(t,n){t instanceof e.TFile&&(this.index.remove(n),this.resolve(t,`rename`))}resolve(t,n=`create`){let r=this.plugin.settings,i=m(r);if(i.length===0)return;let a=this.index.get(t.path);if(!(a&&a.match===`frontmatter`))for(let a of i){let i=r.granularities[a].folder||`/`;if(!t.path.startsWith(i===`/`?``:`${i}/`))continue;let o=f(r,a),s=pe(t,o[0],a),c=window.moment(s,o,!0);if(c.isValid()){let i={filePath:t.path,date:c,granularity:a,match:`filename`};this.index.set(i),n===`create`&&t.stat.size===0&&le(this.app,t,r,i).catch(n=>{console.error(`[Periodic Notes] failed to apply template`,n),new e.Notice(`Periodic Notes: failed to apply template to "${t.path}". See console for details.`)}),this.app.workspace.trigger(`periodic-notes:resolve`,a,t);return}}}getPeriodicNote(t,n){let r=this.index.getByKey(t,n);if(!r)return null;let i=this.app.vault.getAbstractFileByPath(r.filePath);return i instanceof e.TFile?i:(this.index.remove(r.filePath),null)}isPeriodic(e,t){return this.index.has(e,t)}find(e){return this.index.get(e)}findAdjacent(e,t){return this.index.findAdjacent(e,t)}},he=Array.isArray,ge=Array.prototype.indexOf,_e=Array.prototype.includes,ve=Array.from,ye=Object.defineProperty,be=Object.getOwnPropertyDescriptor,xe=Object.getOwnPropertyDescriptors,Se=Object.prototype,Ce=Array.prototype,we=Object.getPrototypeOf,Te=Object.isExtensible;function Ee(e){return typeof e==`function`}var De=()=>{};function Oe(e){for(var t=0;t<e.length;t++)e[t]()}function ke(){var e,t;return{promise:new Promise((n,r)=>{e=n,t=r}),resolve:e,reject:t}}var S=1024,C=2048,Ae=4096,je=8192,Me=16384,Ne=32768,Pe=1<<25,Fe=65536,Ie=1<<19,Le=1<<20,Re=1<<25,ze=65536,Be=1<<21,Ve=1<<22,He=1<<23,Ue=Symbol(`$state`),We=Symbol(`legacy props`),Ge=Symbol(``),Ke=new class extends Error{name=`StaleReactionError`;message="The reaction that called `getAbortSignal()` was re-run or destroyed"},qe=!!globalThis.document?.contentType&&globalThis.document.contentType.includes(`xml`);function Je(e){throw Error(`https://svelte.dev/e/lifecycle_outside_component`)}function Ye(){throw Error(`https://svelte.dev/e/async_derived_orphan`)}function Xe(e,t,n){throw Error(`https://svelte.dev/e/each_key_duplicate`)}function Ze(e){throw Error(`https://svelte.dev/e/effect_in_teardown`)}function Qe(){throw Error(`https://svelte.dev/e/effect_in_unowned_derived`)}function $e(e){throw Error(`https://svelte.dev/e/effect_orphan`)}function et(){throw Error(`https://svelte.dev/e/effect_update_depth_exceeded`)}function tt(e){throw Error(`https://svelte.dev/e/props_invalid_value`)}function nt(){throw Error(`https://svelte.dev/e/set_context_after_init`)}function rt(){throw Error(`https://svelte.dev/e/state_descriptors_fixed`)}function it(){throw Error(`https://svelte.dev/e/state_prototype_fixed`)}function at(){throw Error(`https://svelte.dev/e/state_unsafe_mutation`)}function ot(){throw Error(`https://svelte.dev/e/svelte_boundary_reset_onerror`)}var st={},w=Symbol(),ct=`http://www.w3.org/1999/xhtml`;function lt(e){console.warn(`https://svelte.dev/e/hydration_mismatch`)}function ut(){console.warn(`https://svelte.dev/e/svelte_boundary_reset_noop`)}var T=!1;function dt(e){T=e}var E;function D(e){if(e===null)throw lt(),st;return E=e}function ft(){return D(zn(E))}function O(e){if(T){if(zn(E)!==null)throw lt(),st;E=e}}function pt(e=1){if(T){for(var t=e,n=E;t--;)n=zn(n);E=n}}function mt(e=!0){for(var t=0,n=E;;){if(n.nodeType===8){var r=n.data;if(r===`]`){if(t===0)return n;--t}else (r===`[`||r===`[!`||r[0]===`[`&&!isNaN(Number(r.slice(1))))&&(t+=1)}var i=zn(n);e&&n.remove(),n=i}}function ht(e){if(!e||e.nodeType!==8)throw lt(),st;return e.data}function gt(e){return e===this.v}function _t(e,t){return e==e?e!==t||typeof e==`object`&&!!e||typeof e==`function`:t==t}function vt(e){return!_t(e,this.v)}var yt=!1,bt=!1,k=null;function xt(e){k=e}function St(e){return Dt(`getContext`).get(e)}function Ct(e,t){let n=Dt(`setContext`);if(yt){var r=G.f;!H&&r&32&&!k.i||nt()}return n.set(e,t),t}function wt(e,t=!1,n){k={p:k,i:!1,c:null,e:null,s:e,x:null,r:G,l:bt&&!t?{s:null,u:null,$:[]}:null}}function Tt(e){var t=k,n=t.e;if(n!==null){t.e=null;for(var r of n)Zn(r)}return e!==void 0&&(t.x=e),t.i=!0,k=t.p,e??{}}function Et(){return!bt||k!==null&&k.l===null}function Dt(e){return k===null&&Je(e),k.c??=new Map(Ot(k)||void 0)}function Ot(e){let t=e.p;for(;t!==null;){let e=t.c;if(e!==null)return e;t=t.p}return null}var kt=[];function At(){var e=kt;kt=[],Oe(e)}function jt(e){if(kt.length===0&&!Gt){var t=kt;queueMicrotask(()=>{t===kt&&At()})}kt.push(e)}function Mt(e){var t=G;if(t===null)return H.f|=He,e;if(!(t.f&32768)&&!(t.f&4))throw e;Nt(e,t)}function Nt(e,t){for(;t!==null;){if(t.f&128){if(!(t.f&32768))throw e;try{t.b.error(e);return}catch(t){e=t}}t=t.parent}throw e}var Pt=~(C|Ae|S);function A(e,t){e.f=e.f&Pt|t}function Ft(e){e.f&512||e.deps===null?A(e,S):A(e,Ae)}function It(e){if(e!==null)for(let t of e)!(t.f&2)||!(t.f&65536)||(t.f^=ze,It(t.deps))}function Lt(e,t,n){e.f&2048?t.add(e):e.f&4096&&n.add(e),It(e.deps),A(e,S)}var Rt=[];function zt(e,t=De){let n=null,r=new Set;function i(t){if(_t(e,t)&&(e=t,n)){let t=!Rt.length;for(let t of r)t[1](),Rt.push(t,e);if(t){for(let e=0;e<Rt.length;e+=2)Rt[e][0](Rt[e+1]);Rt.length=0}}}function a(t){i(t(e))}function o(o,s=De){let c=[o,s];return r.add(c),r.size===1&&(n=t(i,a)||De),o(e),()=>{r.delete(c),r.size===0&&n&&(n(),n=null)}}return{set:i,update:a,subscribe:o}}var Bt=!1,Vt=!1;function Ht(e){var t=Vt;try{return Vt=!1,[e(),Vt]}finally{Vt=t}}var Ut=new Set,j=null,M=null,Wt=null,Gt=!1,Kt=!1,qt=null,Jt=null,Yt=0,Xt=1,Zt=class e{id=Xt++;current=new Map;previous=new Map;#e=new Set;#t=new Set;#n=new Map;#r=new Map;#i=null;#a=[];#o=[];#s=new Set;#c=new Set;#l=new Map;is_fork=!1;#u=!1;#d=new Set;#f(){return this.is_fork||this.#r.size>0}#p(){for(let n of this.#d)for(let r of n.#r.keys()){for(var e=!1,t=r;t.parent!==null;){if(this.#l.has(t)){e=!0;break}t=t.parent}if(!e)return!0}return!1}skip_effect(e){this.#l.has(e)||this.#l.set(e,{d:[],m:[]})}unskip_effect(e){var t=this.#l.get(e);if(t){this.#l.delete(e);for(var n of t.d)A(n,C),this.schedule(n);for(n of t.m)A(n,Ae),this.schedule(n)}}#m(){if(Yt++>1e3&&(Ut.delete(this),Qt()),!this.#f()){for(let e of this.#s)this.#c.delete(e),A(e,C),this.schedule(e);for(let e of this.#c)A(e,Ae),this.schedule(e)}let t=this.#a;this.#a=[],this.apply();var n=qt=[],r=[],i=Jt=[];for(let e of t)try{this.#h(e,n,r)}catch(t){throw an(e),t}if(j=null,i.length>0){var a=e.ensure();for(let e of i)a.schedule(e)}if(qt=null,Jt=null,this.#f()||this.#p()){this.#g(r),this.#g(n);for(let[e,t]of this.#l)rn(e,t)}else{this.#n.size===0&&Ut.delete(this),this.#s.clear(),this.#c.clear();for(let e of this.#e)e(this);this.#e.clear(),$t(r),$t(n),this.#i?.resolve()}var o=j;if(this.#a.length>0){let e=o??=this;e.#a.push(...this.#a.filter(t=>!e.#a.includes(t)))}o!==null&&(Ut.add(o),o.#m()),Ut.has(this)||this.#_()}#h(e,t,n){e.f^=S;for(var r=e.first;r!==null;){var i=r.f,a=(i&96)!=0;if(!(a&&i&1024||i&8192||this.#l.has(r))&&r.fn!==null){a?r.f^=S:i&4?t.push(r):yt&&i&16777224?n.push(r):Er(r)&&(i&16&&this.#c.add(r),jr(r));var o=r.first;if(o!==null){r=o;continue}}for(;r!==null;){var s=r.next;if(s!==null){r=s;break}r=r.parent}}}#g(e){for(var t=0;t<e.length;t+=1)Lt(e[t],this.#s,this.#c)}capture(e,t,n=!1){t!==w&&!this.previous.has(e)&&this.previous.set(e,t),e.f&8388608||(this.current.set(e,[e.v,n]),M?.set(e,e.v))}activate(){j=this}deactivate(){j=null,M=null}flush(){try{Kt=!0,j=this,this.#m()}finally{Yt=0,Wt=null,qt=null,Jt=null,Kt=!1,j=null,M=null,wn.clear()}}discard(){for(let e of this.#t)e(this);this.#t.clear(),Ut.delete(this)}register_created_effect(e){this.#o.push(e)}#_(){for(let l of Ut){var e=l.id<this.id,t=[];for(let[r,[i,a]]of this.current){if(l.current.has(r)){var n=l.current.get(r)[0];if(e&&i!==n)l.current.set(r,[i,a]);else continue}t.push(r)}var r=[...l.current.keys()].filter(e=>!this.current.has(e));if(r.length===0)e&&l.discard();else if(t.length>0){l.activate();var i=new Set,a=new Map;for(var o of t)en(o,r,i,a);a=new Map;var s=[...l.current.keys()].filter(e=>this.current.has(e)?this.current.get(e)[0]!==e:!0);for(let e of this.#o)!(e.f&155648)&&tn(e,s,a)&&(e.f&4194320?(A(e,C),l.schedule(e)):l.#s.add(e));if(l.#a.length>0){l.apply();for(var c of l.#a)l.#h(c,[],[]);l.#a=[]}l.deactivate()}}for(let e of Ut)e.#d.has(this)&&(e.#d.delete(this),e.#d.size===0&&!e.#f()&&(e.activate(),e.#m()))}increment(e,t){let n=this.#n.get(t)??0;if(this.#n.set(t,n+1),e){let e=this.#r.get(t)??0;this.#r.set(t,e+1)}}decrement(e,t,n){let r=this.#n.get(t)??0;if(r===1?this.#n.delete(t):this.#n.set(t,r-1),e){let e=this.#r.get(t)??0;e===1?this.#r.delete(t):this.#r.set(t,e-1)}this.#u||n||(this.#u=!0,jt(()=>{this.#u=!1,this.flush()}))}transfer_effects(e,t){for(let t of e)this.#s.add(t);for(let e of t)this.#c.add(e);e.clear(),t.clear()}oncommit(e){this.#e.add(e)}ondiscard(e){this.#t.add(e)}settled(){return(this.#i??=ke()).promise}static ensure(){if(j===null){let t=j=new e;Kt||(Ut.add(j),Gt||jt(()=>{j===t&&t.flush()}))}return j}apply(){if(!yt||!this.is_fork&&Ut.size===1){M=null;return}M=new Map;for(let[e,[t]]of this.current)M.set(e,t);for(let n of Ut)if(!(n===this||n.is_fork)){var e=!1,t=!1;if(n.id<this.id)for(let[r,[,i]]of n.current)i||(e||=this.current.has(r),t||=!this.current.has(r));if(e&&t)this.#d.add(n);else for(let[e,t]of n.previous)M.has(e)||M.set(e,t)}}schedule(e){if(Wt=e,e.b?.is_pending&&e.f&16777228&&!(e.f&32768)){e.b.defer_effect(e);return}for(var t=e;t.parent!==null;){t=t.parent;var n=t.f;if(qt!==null&&t===G&&(yt||(H===null||!(H.f&2))&&!Bt))return;if(n&96){if(!(n&1024))return;t.f^=S}}this.#a.push(t)}};function Qt(){try{et()}catch(e){Nt(e,Wt)}}var N=null;function $t(e){var t=e.length;if(t!==0){for(var n=0;n<t;){var r=e[n++];if(!(r.f&24576)&&Er(r)&&(N=new Set,jr(r),r.deps===null&&r.first===null&&r.nodes===null&&r.teardown===null&&r.ac===null&&cr(r),N?.size>0)){wn.clear();for(let e of N){if(e.f&24576)continue;let t=[e],n=e.parent;for(;n!==null;)N.has(n)&&(N.delete(n),t.push(n)),n=n.parent;for(let e=t.length-1;e>=0;e--){let n=t[e];n.f&24576||jr(n)}}N.clear()}}N=null}}function en(e,t,n,r){if(!n.has(e)&&(n.add(e),e.reactions!==null))for(let i of e.reactions){let e=i.f;e&2?en(i,t,n,r):e&4194320&&!(e&2048)&&tn(i,t,r)&&(A(i,C),nn(i))}}function tn(e,t,n){let r=n.get(e);if(r!==void 0)return r;if(e.deps!==null)for(let r of e.deps){if(_e.call(t,r))return!0;if(r.f&2&&tn(r,t,n))return n.set(r,!0),!0}return n.set(e,!1),!1}function nn(e){j.schedule(e)}function rn(e,t){if(!(e.f&32&&e.f&1024)){e.f&2048?t.d.push(e):e.f&4096&&t.m.push(e),A(e,S);for(var n=e.first;n!==null;)rn(n,t),n=n.next}}function an(e){A(e,S);for(var t=e.first;t!==null;)an(t),t=t.next}function on(e){let t=0,n=En(0),r;return()=>{Jn()&&(X(n),tr(()=>(t===0&&(r=Pr(()=>e(()=>An(n)))),t+=1,()=>{jt(()=>{--t,t===0&&(r?.(),r=void 0,An(n))})})))}}var sn=Fe|Ie;function cn(e,t,n,r){new ln(e,t,n,r)}var ln=class{parent;is_pending=!1;transform_error;#e;#t=T?E:null;#n;#r;#i;#a=null;#o=null;#s=null;#c=null;#l=0;#u=0;#d=!1;#f=new Set;#p=new Set;#m=null;#h=on(()=>(this.#m=En(this.#l),()=>{this.#m=null}));constructor(e,t,n,r){this.#e=e,this.#n=t,this.#r=e=>{var t=G;t.b=this,t.f|=128,n(e)},this.parent=G.b,this.transform_error=r??this.parent?.transform_error??(e=>e),this.#i=rr(()=>{if(T){let e=this.#t;ft();let t=e.data===`[!`;if(e.data.startsWith(`[?`)){let t=JSON.parse(e.data.slice(2));this.#_(t)}else t?this.#v():this.#g()}else this.#y()},sn),T&&(this.#e=E)}#g(){try{this.#a=B(()=>this.#r(this.#e))}catch(e){this.error(e)}}#_(e){let t=this.#n.failed;t&&(this.#s=B(()=>{t(this.#e,()=>e,()=>()=>{})}))}#v(){let e=this.#n.pending;e&&(this.is_pending=!0,this.#o=B(()=>e(this.#e)),jt(()=>{var e=this.#c=document.createDocumentFragment(),t=L();e.append(t),this.#a=this.#x(()=>B(()=>this.#r(t))),this.#u===0&&(this.#e.before(e),this.#c=null,lr(this.#o,()=>{this.#o=null}),this.#b(j))}))}#y(){try{if(this.is_pending=this.has_pending_snippet(),this.#u=0,this.#l=0,this.#a=B(()=>{this.#r(this.#e)}),this.#u>0){var e=this.#c=document.createDocumentFragment();pr(this.#a,e);let t=this.#n.pending;this.#o=B(()=>t(this.#e))}else this.#b(j)}catch(e){this.error(e)}}#b(e){this.is_pending=!1,e.transfer_effects(this.#f,this.#p)}defer_effect(e){Lt(e,this.#f,this.#p)}is_rendered(){return!this.is_pending&&(!this.parent||this.parent.is_rendered())}has_pending_snippet(){return!!this.#n.pending}#x(e){var t=G,n=H,r=k;vr(this.#i),W(this.#i),xt(this.#i.ctx);try{return Zt.ensure(),e()}catch(e){return Mt(e),null}finally{vr(t),W(n),xt(r)}}#S(e,t){if(!this.has_pending_snippet()){this.parent&&this.parent.#S(e,t);return}this.#u+=e,this.#u===0&&(this.#b(t),this.#o&&lr(this.#o,()=>{this.#o=null}),this.#c&&=(this.#e.before(this.#c),null))}update_pending_count(e,t){this.#S(e,t),this.#l+=e,!(!this.#m||this.#d)&&(this.#d=!0,jt(()=>{this.#d=!1,this.#m&&On(this.#m,this.#l)}))}get_effect_pending(){return this.#h(),X(this.#m)}error(e){var t=this.#n.onerror;let n=this.#n.failed;if(!t&&!n)throw e;this.#a&&=(V(this.#a),null),this.#o&&=(V(this.#o),null),this.#s&&=(V(this.#s),null),T&&(D(this.#t),pt(),D(mt()));var r=!1,i=!1;let a=()=>{if(r){ut();return}r=!0,i&&ot(),this.#s!==null&&lr(this.#s,()=>{this.#s=null}),this.#x(()=>{this.#y()})},o=e=>{try{i=!0,t?.(e,a),i=!1}catch(e){Nt(e,this.#i&&this.#i.parent)}n&&(this.#s=this.#x(()=>{try{return B(()=>{var t=G;t.b=this,t.f|=128,n(this.#e,()=>e,()=>a)})}catch(e){return Nt(e,this.#i.parent),null}}))};jt(()=>{var t;try{t=this.transform_error(e)}catch(e){Nt(e,this.#i&&this.#i.parent);return}typeof t==`object`&&t&&typeof t.then==`function`?t.then(o,e=>Nt(e,this.#i&&this.#i.parent)):o(t)})}};function un(e,t,n,r){let i=Et()?mn:gn;var a=e.filter(e=>!e.settled);if(n.length===0&&a.length===0){r(t.map(i));return}var o=G,s=dn(),c=a.length===1?a[0].promise:a.length>1?Promise.all(a.map(e=>e.promise)):null;function l(e){s();try{r(e)}catch(e){o.f&16384||Nt(e,o)}fn()}if(n.length===0){c.then(()=>l(t.map(i)));return}var u=pn();function d(){Promise.all(n.map(e=>hn(e))).then(e=>l([...t.map(i),...e])).catch(e=>Nt(e,o)).finally(()=>u())}c?c.then(()=>{s(),d(),fn()}):d()}function dn(){var e=G,t=H,n=k,r=j;return function(i=!0){vr(e),W(t),xt(n),i&&!(e.f&16384)&&(r?.activate(),r?.apply())}}function fn(e=!0){vr(null),W(null),xt(null),e&&j?.deactivate()}function pn(){var e=G,t=e.b,n=j,r=t.is_rendered();return t.update_pending_count(1,n),n.increment(r,e),(i=!1)=>{t.update_pending_count(-1,n),n.decrement(r,e,i)}}function mn(e){var t=2|C,n=H!==null&&H.f&2?H:null;return G!==null&&(G.f|=Ie),{ctx:k,deps:null,effects:null,equals:gt,f:t,fn:e,reactions:null,rv:0,v:w,wv:0,parent:n??G,ac:null}}function hn(e,t,n){let r=G;r===null&&Ye();var i=void 0,a=En(w),o=!H,s=new Map;return er(()=>{var t=G,n=ke();i=n.promise;try{Promise.resolve(e()).then(n.resolve,n.reject).finally(fn)}catch(e){n.reject(e),fn()}var c=j;if(o){if(t.f&32768)var l=pn();if(r.b.is_rendered())s.get(c)?.reject(Ke),s.delete(c);else{for(let e of s.values())e.reject(Ke);s.clear()}s.set(c,n)}let u=(e,n=void 0)=>{if(l&&l(n===Ke),!(n===Ke||t.f&16384)){if(c.activate(),n)a.f|=He,On(a,n);else{a.f&8388608&&(a.f^=He),On(a,e);for(let[e,t]of s){if(s.delete(e),e===c)break;t.reject(Ke)}}c.deactivate()}};n.promise.then(u,e=>u(null,e||`unknown`))}),Yn(()=>{for(let e of s.values())e.reject(Ke)}),new Promise(e=>{function t(n){function r(){n===i?e(a):t(i)}n.then(r,r)}t(i)})}function P(e){let t=mn(e);return yt||yr(t),t}function gn(e){let t=mn(e);return t.equals=vt,t}function _n(e){var t=e.effects;if(t!==null){e.effects=null;for(var n=0;n<t.length;n+=1)V(t[n])}}function vn(e){for(var t=e.parent;t!==null;){if(!(t.f&2))return t.f&16384?null:t;t=t.parent}return null}function yn(e){var t,n=G;vr(vn(e));try{e.f&=~ze,_n(e),t=Or(e)}finally{vr(n)}return t}function bn(e){var t=e.v,n=yn(e);if(!e.equals(n)&&(e.wv=Tr(),(!j?.is_fork||e.deps===null)&&(e.v=n,j?.capture(e,t,!0),e.deps===null))){A(e,S);return}gr||(M===null?Ft(e):(Jn()||j?.is_fork)&&M.set(e,n))}function xn(e){if(e.effects!==null)for(let t of e.effects)(t.teardown||t.ac)&&(t.teardown?.(),t.ac?.abort(Ke),t.teardown=De,t.ac=null,Ar(t,0),ar(t))}function Sn(e){if(e.effects!==null)for(let t of e.effects)t.teardown&&jr(t)}var Cn=new Set,wn=new Map,Tn=!1;function En(e,t){return{f:0,v:e,reactions:null,equals:gt,rv:0,wv:0}}function F(e,t){let n=En(e,t);return yr(n),n}function Dn(e,t=!1,n=!0){let r=En(e);return t||(r.equals=vt),bt&&n&&k!==null&&k.l!==null&&(k.l.s??=[]).push(r),r}function I(e,t,n=!1){return H!==null&&(!U||H.f&131072)&&Et()&&H.f&4325394&&(K===null||!_e.call(K,e))&&at(),On(e,n?Mn(t):t,Jt)}function On(e,t,n=null){if(!e.equals(t)){var r=e.v;gr?wn.set(e,t):wn.set(e,r),e.v=t;var i=Zt.ensure();if(i.capture(e,r),e.f&2){let t=e;e.f&2048&&yn(t),M===null&&Ft(t)}e.wv=Tr(),jn(e,C,n),Et()&&G!==null&&G.f&1024&&!(G.f&96)&&(Y===null?br([e]):Y.push(e)),!i.is_fork&&Cn.size>0&&!Tn&&kn()}return t}function kn(){Tn=!1;for(let e of Cn)e.f&1024&&A(e,Ae),Er(e)&&jr(e);Cn.clear()}function An(e){I(e,e.v+1)}function jn(e,t,n){var r=e.reactions;if(r!==null)for(var i=Et(),a=r.length,o=0;o<a;o++){var s=r[o],c=s.f;if(!(!i&&s===G)){var l=(c&C)===0;if(l&&A(s,t),c&2){var u=s;M?.delete(u),c&65536||(c&512&&(s.f|=ze),jn(u,Ae,n))}else if(l){var d=s;c&16&&N!==null&&N.add(d),n===null?nn(d):n.push(d)}}}}function Mn(e){if(typeof e!=`object`||!e||Ue in e)return e;let t=we(e);if(t!==Se&&t!==Ce)return e;var n=new Map,r=he(e),i=F(0),a=null,o=Cr,s=e=>{if(Cr===o)return e();var t=H,n=Cr;W(null),wr(o);var r=e();return W(t),wr(n),r};return r&&n.set(`length`,F(e.length,a)),new Proxy(e,{defineProperty(e,t,r){(!(`value`in r)||r.configurable===!1||r.enumerable===!1||r.writable===!1)&&rt();var i=n.get(t);return i===void 0?s(()=>{var e=F(r.value,a);return n.set(t,e),e}):I(i,r.value,!0),!0},deleteProperty(e,t){var r=n.get(t);if(r===void 0){if(t in e){let e=s(()=>F(w,a));n.set(t,e),An(i)}}else I(r,w),An(i);return!0},get(t,r,i){if(r===Ue)return e;var o=n.get(r),c=r in t;if(o===void 0&&(!c||be(t,r)?.writable)&&(o=s(()=>F(Mn(c?t[r]:w),a)),n.set(r,o)),o!==void 0){var l=X(o);return l===w?void 0:l}return Reflect.get(t,r,i)},getOwnPropertyDescriptor(e,t){var r=Reflect.getOwnPropertyDescriptor(e,t);if(r&&`value`in r){var i=n.get(t);i&&(r.value=X(i))}else if(r===void 0){var a=n.get(t),o=a?.v;if(a!==void 0&&o!==w)return{enumerable:!0,configurable:!0,value:o,writable:!0}}return r},has(e,t){if(t===Ue)return!0;var r=n.get(t),i=r!==void 0&&r.v!==w||Reflect.has(e,t);return(r!==void 0||G!==null&&(!i||be(e,t)?.writable))&&(r===void 0&&(r=s(()=>F(i?Mn(e[t]):w,a)),n.set(t,r)),X(r)===w)?!1:i},set(e,t,o,c){var l=n.get(t),u=t in e;if(r&&t===`length`)for(var d=o;d<l.v;d+=1){var f=n.get(d+``);f===void 0?d in e&&(f=s(()=>F(w,a)),n.set(d+``,f)):I(f,w)}if(l===void 0)(!u||be(e,t)?.writable)&&(l=s(()=>F(void 0,a)),I(l,Mn(o)),n.set(t,l));else{u=l.v!==w;var p=s(()=>Mn(o));I(l,p)}var m=Reflect.getOwnPropertyDescriptor(e,t);if(m?.set&&m.set.call(c,o),!u){if(r&&typeof t==`string`){var h=n.get(`length`),g=Number(t);Number.isInteger(g)&&g>=h.v&&I(h,g+1)}An(i)}return!0},ownKeys(e){X(i);var t=Reflect.ownKeys(e).filter(e=>{var t=n.get(e);return t===void 0||t.v!==w});for(var[r,a]of n)a.v!==w&&!(r in e)&&t.push(r);return t},setPrototypeOf(){it()}})}var Nn,Pn,Fn,In;function Ln(){if(Nn===void 0){Nn=window,Pn=/Firefox/.test(navigator.userAgent);var e=Element.prototype,t=Node.prototype,n=Text.prototype;Fn=be(t,`firstChild`).get,In=be(t,`nextSibling`).get,Te(e)&&(e.__click=void 0,e.__className=void 0,e.__attributes=null,e.__style=void 0,e.__e=void 0),Te(n)&&(n.__t=void 0)}}function L(e=``){return document.createTextNode(e)}function Rn(e){return Fn.call(e)}function zn(e){return In.call(e)}function R(e,t){if(!T)return Rn(e);var n=Rn(E);if(n===null)n=E.appendChild(L());else if(t&&n.nodeType!==3){var r=L();return n?.before(r),D(r),r}return t&&Un(n),D(n),n}function z(e,t=1,n=!1){let r=T?E:e;for(var i;t--;)i=r,r=zn(r);if(!T)return r;if(n){if(r?.nodeType!==3){var a=L();return r===null?i?.after(a):r.before(a),D(a),a}Un(r)}return D(r),r}function Bn(e){e.textContent=``}function Vn(){return!yt||N!==null?!1:(G.f&Ne)!==0}function Hn(e,t,n){let r=n?{is:n}:void 0;return document.createElementNS(t??`http://www.w3.org/1999/xhtml`,e,r)}function Un(e){if(e.nodeValue.length<65536)return;let t=e.nextSibling;for(;t!==null&&t.nodeType===3;)t.remove(),e.nodeValue+=t.nodeValue,t=e.nextSibling}function Wn(e){var t=H,n=G;W(null),vr(null);try{return e()}finally{W(t),vr(n)}}function Gn(e){G===null&&(H===null&&$e(e),Qe()),gr&&Ze(e)}function Kn(e,t){var n=t.last;n===null?t.last=t.first=e:(n.next=e,e.prev=n,t.last=e)}function qn(e,t){var n=G;n!==null&&n.f&8192&&(e|=je);var r={ctx:k,deps:null,nodes:null,f:e|C|512,first:null,fn:t,last:null,next:null,parent:n,b:n&&n.b,prev:null,teardown:null,wv:0,ac:null};j?.register_created_effect(r);var i=r;if(e&4)qt===null?Zt.ensure().schedule(r):qt.push(r);else if(t!==null){try{jr(r)}catch(e){throw V(r),e}i.deps===null&&i.teardown===null&&i.nodes===null&&i.first===i.last&&!(i.f&524288)&&(i=i.first,e&16&&e&65536&&i!==null&&(i.f|=Fe))}if(i!==null&&(i.parent=n,n!==null&&Kn(i,n),H!==null&&H.f&2&&!(e&64))){var a=H;(a.effects??=[]).push(i)}return r}function Jn(){return H!==null&&!U}function Yn(e){let t=qn(8,null);return A(t,S),t.teardown=e,t}function Xn(e){Gn(`$effect`);var t=G.f;if(!H&&t&32&&!(t&32768)){var n=k;(n.e??=[]).push(e)}else return Zn(e)}function Zn(e){return qn(4|Le,e)}function Qn(e){Zt.ensure();let t=qn(64|Ie,e);return(e={})=>new Promise(n=>{e.outro?lr(t,()=>{V(t),n(void 0)}):(V(t),n(void 0))})}function $n(e){return qn(4,e)}function er(e){return qn(Ve|Ie,e)}function tr(e,t=0){return qn(8|t,e)}function nr(e,t=[],n=[],r=[]){un(r,t,n,t=>{qn(8,()=>e(...t.map(X)))})}function rr(e,t=0){return qn(16|t,e)}function B(e){return qn(32|Ie,e)}function ir(e){var t=e.teardown;if(t!==null){let e=gr,n=H;_r(!0),W(null);try{t.call(null)}finally{_r(e),W(n)}}}function ar(e,t=!1){var n=e.first;for(e.first=e.last=null;n!==null;){let e=n.ac;e!==null&&Wn(()=>{e.abort(Ke)});var r=n.next;n.f&64?n.parent=null:V(n,t),n=r}}function or(e){for(var t=e.first;t!==null;){var n=t.next;t.f&32||V(t),t=n}}function V(e,t=!0){var n=!1;(t||e.f&262144)&&e.nodes!==null&&e.nodes.end!==null&&(sr(e.nodes.start,e.nodes.end),n=!0),A(e,Pe),ar(e,t&&!n),Ar(e,0);var r=e.nodes&&e.nodes.t;if(r!==null)for(let e of r)e.stop();ir(e),e.f^=Pe,e.f|=Me;var i=e.parent;i!==null&&i.first!==null&&cr(e),e.next=e.prev=e.teardown=e.ctx=e.deps=e.fn=e.nodes=e.ac=e.b=null}function sr(e,t){for(;e!==null;){var n=e===t?null:zn(e);e.remove(),e=n}}function cr(e){var t=e.parent,n=e.prev,r=e.next;n!==null&&(n.next=r),r!==null&&(r.prev=n),t!==null&&(t.first===e&&(t.first=r),t.last===e&&(t.last=n))}function lr(e,t,n=!0){var r=[];ur(e,r,!0);var i=()=>{n&&V(e),t&&t()},a=r.length;if(a>0){var o=()=>--a||i();for(var s of r)s.out(o)}else i()}function ur(e,t,n){if(!(e.f&8192)){e.f^=je;var r=e.nodes&&e.nodes.t;if(r!==null)for(let e of r)(e.is_global||n)&&t.push(e);for(var i=e.first;i!==null;){var a=i.next,o=(i.f&65536)!=0||(i.f&32)!=0&&(e.f&16)!=0;ur(i,t,o?n:!1),i=a}}}function dr(e){fr(e,!0)}function fr(e,t){if(e.f&8192){e.f^=je,e.f&1024||(A(e,C),Zt.ensure().schedule(e));for(var n=e.first;n!==null;){var r=n.next,i=(n.f&65536)!=0||(n.f&32)!=0;fr(n,i?t:!1),n=r}var a=e.nodes&&e.nodes.t;if(a!==null)for(let e of a)(e.is_global||t)&&e.in()}}function pr(e,t){if(e.nodes)for(var n=e.nodes.start,r=e.nodes.end;n!==null;){var i=n===r?null:zn(n);t.append(n),n=i}}var mr=null,hr=!1,gr=!1;function _r(e){gr=e}var H=null,U=!1;function W(e){H=e}var G=null;function vr(e){G=e}var K=null;function yr(e){H!==null&&(!yt||H.f&2)&&(K===null?K=[e]:K.push(e))}var q=null,J=0,Y=null;function br(e){Y=e}var xr=1,Sr=0,Cr=Sr;function wr(e){Cr=e}function Tr(){return++xr}function Er(e){var t=e.f;if(t&2048)return!0;if(t&2&&(e.f&=~ze),t&4096){for(var n=e.deps,r=n.length,i=0;i<r;i++){var a=n[i];if(Er(a)&&bn(a),a.wv>e.wv)return!0}t&512&&M===null&&A(e,S)}return!1}function Dr(e,t,n=!0){var r=e.reactions;if(r!==null&&!(!yt&&K!==null&&_e.call(K,e)))for(var i=0;i<r.length;i++){var a=r[i];a.f&2?Dr(a,t,!1):t===a&&(n?A(a,C):a.f&1024&&A(a,Ae),nn(a))}}function Or(e){var t=q,n=J,r=Y,i=H,a=K,o=k,s=U,c=Cr,l=e.f;q=null,J=0,Y=null,H=l&96?null:e,K=null,xt(e.ctx),U=!1,Cr=++Sr,e.ac!==null&&(Wn(()=>{e.ac.abort(Ke)}),e.ac=null);try{e.f|=Be;var u=e.fn,d=u();e.f|=Ne;var f=e.deps,p=j?.is_fork;if(q!==null){var m;if(p||Ar(e,J),f!==null&&J>0)for(f.length=J+q.length,m=0;m<q.length;m++)f[J+m]=q[m];else e.deps=f=q;if(Jn()&&e.f&512)for(m=J;m<f.length;m++)(f[m].reactions??=[]).push(e)}else !p&&f!==null&&J<f.length&&(Ar(e,J),f.length=J);if(Et()&&Y!==null&&!U&&f!==null&&!(e.f&6146))for(m=0;m<Y.length;m++)Dr(Y[m],e);if(i!==null&&i!==e){if(Sr++,i.deps!==null)for(let e=0;e<n;e+=1)i.deps[e].rv=Sr;if(t!==null)for(let e of t)e.rv=Sr;Y!==null&&(r===null?r=Y:r.push(...Y))}return e.f&8388608&&(e.f^=He),d}catch(e){return Mt(e)}finally{e.f^=Be,q=t,J=n,Y=r,H=i,K=a,xt(o),U=s,Cr=c}}function kr(e,t){let n=t.reactions;if(n!==null){var r=ge.call(n,e);if(r!==-1){var i=n.length-1;i===0?n=t.reactions=null:(n[r]=n[i],n.pop())}}if(n===null&&t.f&2&&(q===null||!_e.call(q,t))){var a=t;a.f&512&&(a.f^=512,a.f&=~ze),Ft(a),xn(a),Ar(a,0)}}function Ar(e,t){var n=e.deps;if(n!==null)for(var r=t;r<n.length;r++)kr(e,n[r])}function jr(e){var t=e.f;if(!(t&16384)){A(e,S);var n=G,r=hr;G=e,hr=!0;try{t&16777232?or(e):ar(e),ir(e);var i=Or(e);e.teardown=typeof i==`function`?i:null,e.wv=xr}finally{hr=r,G=n}}}function X(e){var t=(e.f&2)!=0;if(mr?.add(e),H!==null&&!U&&!(G!==null&&G.f&16384)&&(K===null||!_e.call(K,e))){var n=H.deps;if(H.f&2097152)e.rv<Sr&&(e.rv=Sr,q===null&&n!==null&&n[J]===e?J++:q===null?q=[e]:q.push(e));else{(H.deps??=[]).push(e);var r=e.reactions;r===null?e.reactions=[H]:_e.call(r,H)||r.push(H)}}if(gr&&wn.has(e))return wn.get(e);if(t){var i=e;if(gr){var a=i.v;return(!(i.f&1024)&&i.reactions!==null||Nr(i))&&(a=yn(i)),wn.set(i,a),a}var o=(i.f&512)==0&&!U&&H!==null&&(hr||(H.f&512)!=0),s=(i.f&Ne)===0;Er(i)&&(o&&(i.f|=512),bn(i)),o&&!s&&(Sn(i),Mr(i))}if(M?.has(e))return M.get(e);if(e.f&8388608)throw e.v;return e.v}function Mr(e){if(e.f|=512,e.deps!==null)for(let t of e.deps)(t.reactions??=[]).push(e),t.f&2&&!(t.f&512)&&(Sn(t),Mr(t))}function Nr(e){if(e.v===w)return!0;if(e.deps===null)return!1;for(let t of e.deps)if(wn.has(t)||t.f&2&&Nr(t))return!0;return!1}function Pr(e){var t=U;try{return U=!0,e()}finally{U=t}}[...`allowfullscreen.async.autofocus.autoplay.checked.controls.default.disabled.formnovalidate.indeterminate.inert.ismap.loop.multiple.muted.nomodule.novalidate.open.playsinline.readonly.required.reversed.seamless.selected.webkitdirectory.defer.disablepictureinpicture.disableremoteplayback`.split(`.`)];var Fr=[`touchstart`,`touchmove`];function Ir(e){return Fr.includes(e)}var Lr=Symbol(`events`),Rr=new Set,zr=new Set;function Br(e,t,n,r={}){function i(e){if(r.capture||Wr.call(t,e),!e.cancelBubble)return Wn(()=>n?.call(this,e))}return e.startsWith(`pointer`)||e.startsWith(`touch`)||e===`wheel`?jt(()=>{t.addEventListener(e,i,r)}):t.addEventListener(e,i,r),i}function Vr(e,t,n,r,i){var a={capture:r,passive:i},o=Br(e,t,n,a);(t===document.body||t===window||t===document||t instanceof HTMLMediaElement)&&Yn(()=>{t.removeEventListener(e,o,a)})}function Z(e,t,n){(t[Lr]??={})[e]=n}function Hr(e){for(var t=0;t<e.length;t++)Rr.add(e[t]);for(var n of zr)n(e)}var Ur=null;function Wr(e){var t=this,n=t.ownerDocument,r=e.type,i=e.composedPath?.()||[],a=i[0]||e.target;Ur=e;var o=0,s=Ur===e&&e[Lr];if(s){var c=i.indexOf(s);if(c!==-1&&(t===document||t===window)){e[Lr]=t;return}var l=i.indexOf(t);if(l===-1)return;c<=l&&(o=c)}if(a=i[o]||e.target,a!==t){ye(e,`currentTarget`,{configurable:!0,get(){return a||n}});var u=H,d=G;W(null),vr(null);try{for(var f,p=[];a!==null;){var m=a.assignedSlot||a.parentNode||a.host||null;try{var h=a[Lr]?.[r];h!=null&&(!a.disabled||e.target===a)&&h.call(a,e)}catch(e){f?p.push(e):f=e}if(e.cancelBubble||m===t||m===null)break;a=m}if(f){for(let e of p)queueMicrotask(()=>{throw e});throw f}}finally{e[Lr]=t,delete e.currentTarget,W(u),vr(d)}}}var Gr=globalThis?.window?.trustedTypes&&globalThis.window.trustedTypes.createPolicy(`svelte-trusted-html`,{createHTML:e=>e});function Kr(e){return Gr?.createHTML(e)??e}function qr(e){var t=Hn(`template`);return t.innerHTML=Kr(e.replaceAll(`<!>`,`<!---->`)),t.content}function Jr(e,t){var n=G;n.nodes===null&&(n.nodes={start:e,end:t,a:null,t:null})}function Q(e,t){var n=(t&1)!=0,r=(t&2)!=0,i,a=!e.startsWith(`<!>`);return()=>{if(T)return Jr(E,null),E;i===void 0&&(i=qr(a?e:`<!>`+e),n||(i=Rn(i)));var t=r||Pn?document.importNode(i,!0):i.cloneNode(!0);if(n){var o=Rn(t),s=t.lastChild;Jr(o,s)}else Jr(t,t);return t}}function $(e,t){if(T){var n=G;(!(n.f&32768)||n.nodes.end===null)&&(n.nodes.end=E),ft();return}e!==null&&e.before(t)}function Yr(e,t){var n=t==null?``:typeof t==`object`?`${t}`:t;n!==(e.__t??=e.nodeValue)&&(e.__t=n,e.nodeValue=`${n}`)}function Xr(e,t){return Qr(e,t)}var Zr=new Map;function Qr(e,{target:t,anchor:n,props:r={},events:i,context:a,intro:o=!0,transformError:s}){Ln();var c=void 0,l=Qn(()=>{var o=n??t.appendChild(L());cn(o,{pending:()=>{}},t=>{wt({});var n=k;if(a&&(n.c=a),i&&(r.$$events=i),T&&Jr(t,null),c=e(t,r)||{},T&&(G.nodes.end=E,E===null||E.nodeType!==8||E.data!==`]`))throw lt(),st;Tt()},s);var l=new Set,u=e=>{for(var n=0;n<e.length;n++){var r=e[n];if(!l.has(r)){l.add(r);var i=Ir(r);for(let e of[t,document]){var a=Zr.get(e);a===void 0&&(a=new Map,Zr.set(e,a));var o=a.get(r);o===void 0?(e.addEventListener(r,Wr,{passive:i}),a.set(r,1)):a.set(r,o+1)}}}};return u(ve(Rr)),zr.add(u),()=>{for(var e of l)for(let n of[t,document]){var r=Zr.get(n),i=r.get(e);--i==0?(n.removeEventListener(e,Wr),r.delete(e),r.size===0&&Zr.delete(n)):r.set(e,i)}zr.delete(u),o!==n&&o.parentNode?.removeChild(o)}});return $r.set(c,l),c}var $r=new WeakMap;function ei(e,t){let n=$r.get(e);return n?($r.delete(e),n(t)):Promise.resolve()}var ti=class{anchor;#e=new Map;#t=new Map;#n=new Map;#r=new Set;#i=!0;constructor(e,t=!0){this.anchor=e,this.#i=t}#a=e=>{if(this.#e.has(e)){var t=this.#e.get(e),n=this.#t.get(t);if(n)dr(n),this.#r.delete(t);else{var r=this.#n.get(t);r&&(this.#t.set(t,r.effect),this.#n.delete(t),r.fragment.lastChild.remove(),this.anchor.before(r.fragment),n=r.effect)}for(let[t,n]of this.#e){if(this.#e.delete(t),t===e)break;let r=this.#n.get(n);r&&(V(r.effect),this.#n.delete(n))}for(let[e,r]of this.#t){if(e===t||this.#r.has(e))continue;let i=()=>{if(Array.from(this.#e.values()).includes(e)){var t=document.createDocumentFragment();pr(r,t),t.append(L()),this.#n.set(e,{effect:r,fragment:t})}else V(r);this.#r.delete(e),this.#t.delete(e)};this.#i||!n?(this.#r.add(e),lr(r,i,!1)):i()}}};#o=e=>{this.#e.delete(e);let t=Array.from(this.#e.values());for(let[e,n]of this.#n)t.includes(e)||(V(n.effect),this.#n.delete(e))};ensure(e,t){var n=j,r=Vn();if(t&&!this.#t.has(e)&&!this.#n.has(e))if(r){var i=document.createDocumentFragment(),a=L();i.append(a),this.#n.set(e,{effect:B(()=>t(a)),fragment:i})}else this.#t.set(e,B(()=>t(this.anchor)));if(this.#e.set(n,e),r){for(let[t,r]of this.#t)t===e?n.unskip_effect(r):n.skip_effect(r);for(let[t,r]of this.#n)t===e?n.unskip_effect(r.effect):n.skip_effect(r.effect);n.oncommit(this.#a),n.ondiscard(this.#o)}else T&&(this.anchor=E),this.#a(n)}};function ni(e,t,n=!1){var r;T&&(r=E,ft());var i=new ti(e),a=n?Fe:0;function o(e,t){if(T){var n=ht(r);if(e!==parseInt(n.substring(1))){var a=mt();D(a),i.anchor=a,dt(!1),i.ensure(e,t),dt(!0);return}}i.ensure(e,t)}rr(()=>{var e=!1;t((t,n=0)=>{e=!0,o(n,t)}),e||o(-1,null)},a)}function ri(e,t){return t}function ii(e,t,n){for(var r=[],i=t.length,a,o=t.length,s=0;s<i;s++){let n=t[s];lr(n,()=>{if(a){if(a.pending.delete(n),a.done.add(n),a.pending.size===0){var t=e.outrogroups;ai(e,ve(a.done)),t.delete(a),t.size===0&&(e.outrogroups=null)}}else --o},!1)}if(o===0){var c=r.length===0&&n!==null;if(c){var l=n,u=l.parentNode;Bn(u),u.append(l),e.items.clear()}ai(e,t,!c)}else a={pending:new Set(t),done:new Set},(e.outrogroups??=new Set).add(a)}function ai(e,t,n=!0){var r;if(e.pending.size>0){r=new Set;for(let t of e.pending.values())for(let n of t)r.add(e.items.get(n).e)}for(var i=0;i<t.length;i++){var a=t[i];r?.has(a)?(a.f|=Re,pr(a,document.createDocumentFragment())):V(t[i],n)}}var oi;function si(e,t,n,r,i,a=null){var o=e,s=new Map;if(t&4){var c=e;o=T?D(Rn(c)):c.appendChild(L())}T&&ft();var l=null,u=gn(()=>{var e=n();return he(e)?e:e==null?[]:ve(e)}),d,f=new Map,p=!0;function m(e){g.effect.f&16384||(g.pending.delete(e),g.fallback=l,li(g,d,o,t,r),l!==null&&(d.length===0?l.f&33554432?(l.f^=Re,di(l,null,o)):dr(l):lr(l,()=>{l=null})))}function h(e){g.pending.delete(e)}var g={effect:rr(()=>{d=X(u);var e=d.length;let c=!1;T&&ht(o)===`[!`!=(e===0)&&(o=mt(),D(o),dt(!1),c=!0);for(var g=new Set,_=j,v=Vn(),y=0;y<e;y+=1){T&&E.nodeType===8&&E.data===`]`&&(o=E,c=!0,dt(!1));var b=d[y],ee=r(b,y),x=p?null:s.get(ee);x?(x.v&&On(x.v,b),x.i&&On(x.i,y),v&&_.unskip_effect(x.e)):(x=ui(s,p?o:oi??=L(),b,ee,y,i,t,n),p||(x.e.f|=Re),s.set(ee,x)),g.add(ee)}if(e===0&&a&&!l&&(p?l=B(()=>a(o)):(l=B(()=>a(oi??=L())),l.f|=Re)),e>g.size&&Xe(``,``,``),T&&e>0&&D(mt()),!p)if(f.set(_,g),v){for(let[e,t]of s)g.has(e)||_.skip_effect(t.e);_.oncommit(m),_.ondiscard(h)}else m(_);c&&dt(!0),X(u)}),flags:t,items:s,pending:f,outrogroups:null,fallback:l};p=!1,T&&(o=E)}function ci(e){for(;e!==null&&!(e.f&32);)e=e.next;return e}function li(e,t,n,r,i){var a=(r&8)!=0,o=t.length,s=e.items,c=ci(e.effect.first),l,u=null,d,f=[],p=[],m,h,g,_;if(a)for(_=0;_<o;_+=1)m=t[_],h=i(m,_),g=s.get(h).e,g.f&33554432||(g.nodes?.a?.measure(),(d??=new Set).add(g));for(_=0;_<o;_+=1){if(m=t[_],h=i(m,_),g=s.get(h).e,e.outrogroups!==null)for(let t of e.outrogroups)t.pending.delete(g),t.done.delete(g);if(g.f&8192&&(dr(g),a&&(g.nodes?.a?.unfix(),(d??=new Set).delete(g))),g.f&33554432)if(g.f^=Re,g===c)di(g,null,n);else{var v=u?u.next:c;g===e.effect.last&&(e.effect.last=g.prev),g.prev&&(g.prev.next=g.next),g.next&&(g.next.prev=g.prev),fi(e,u,g),fi(e,g,v),di(g,v,n),u=g,f=[],p=[],c=ci(u.next);continue}if(g!==c){if(l!==void 0&&l.has(g)){if(f.length<p.length){var y=p[0],b;u=y.prev;var ee=f[0],x=f[f.length-1];for(b=0;b<f.length;b+=1)di(f[b],y,n);for(b=0;b<p.length;b+=1)l.delete(p[b]);fi(e,ee.prev,x.next),fi(e,u,ee),fi(e,x,y),c=y,u=x,--_,f=[],p=[]}else l.delete(g),di(g,c,n),fi(e,g.prev,g.next),fi(e,g,u===null?e.effect.first:u.next),fi(e,u,g),u=g;continue}for(f=[],p=[];c!==null&&c!==g;)(l??=new Set).add(c),p.push(c),c=ci(c.next);if(c===null)continue}g.f&33554432||f.push(g),u=g,c=ci(g.next)}if(e.outrogroups!==null){for(let t of e.outrogroups)t.pending.size===0&&(ai(e,ve(t.done)),e.outrogroups?.delete(t));e.outrogroups.size===0&&(e.outrogroups=null)}if(c!==null||l!==void 0){var te=[];if(l!==void 0)for(g of l)g.f&8192||te.push(g);for(;c!==null;)!(c.f&8192)&&c!==e.fallback&&te.push(c),c=ci(c.next);var ne=te.length;if(ne>0){var re=r&4&&o===0?n:null;if(a){for(_=0;_<ne;_+=1)te[_].nodes?.a?.measure();for(_=0;_<ne;_+=1)te[_].nodes?.a?.fix()}ii(e,te,re)}}a&&jt(()=>{if(d!==void 0)for(g of d)g.nodes?.a?.apply()})}function ui(e,t,n,r,i,a,o,s){var c=o&1?o&16?En(n):Dn(n,!1,!1):null,l=o&2?En(i):null;return{v:c,i:l,e:B(()=>(a(t,c??n,l??i,s),()=>{e.delete(r)}))}}function di(e,t,n){if(e.nodes)for(var r=e.nodes.start,i=e.nodes.end,a=t&&!(t.f&33554432)?t.nodes.start:n;r!==null;){var o=zn(r);if(a.before(r),r===i)return;r=o}}function fi(e,t,n){t===null?e.effect.first=n:t.next=n,n===null?e.effect.last=t:n.prev=t}function pi(e,t){$n(()=>{var n=e.getRootNode(),r=n.host?n:n.head??n.ownerDocument.head;if(!r.querySelector(`#`+t.hash)){let e=Hn(`style`);e.id=t.hash,e.textContent=t.code,r.appendChild(e)}})}var mi=[...` 	
-\r\f\xA0\v﻿`];function hi(e,t,n){var r=e==null?``:``+e;if(t&&(r=r?r+` `+t:t),n){for(var i of Object.keys(n))if(n[i])r=r?r+` `+i:i;else if(r.length)for(var a=i.length,o=0;(o=r.indexOf(i,o))>=0;){var s=o+a;(o===0||mi.includes(r[o-1]))&&(s===r.length||mi.includes(r[s]))?r=(o===0?``:r.substring(0,o))+r.substring(s+1):o=s}}return r===``?null:r}function gi(e,t,n,r,i,a){var o=e.__className;if(T||o!==n||o===void 0){var s=hi(n,r,a);(!T||s!==e.getAttribute(`class`))&&(s==null?e.removeAttribute(`class`):t?e.className=s:e.setAttribute(`class`,s)),e.__className=n}else if(a&&i!==a)for(var c in a){var l=!!a[c];(i==null||l!==!!i[c])&&e.classList.toggle(c,l)}return a}var _i=Symbol(`is custom element`),vi=Symbol(`is html`),yi=qe?`link`:`LINK`;function bi(e,t,n,r){var i=xi(e);T&&(i[t]=e.getAttribute(t),t===`src`||t===`srcset`||t===`href`&&e.nodeName===yi)||i[t]!==(i[t]=n)&&(t===`loading`&&(e[Ge]=n),n==null?e.removeAttribute(t):typeof n!=`string`&&Ci(e).includes(t)?e[t]=n:e.setAttribute(t,n))}function xi(e){return e.__attributes??={[_i]:e.nodeName.includes(`-`),[vi]:e.namespaceURI===ct}}var Si=new Map;function Ci(e){var t=e.getAttribute(`is`)||e.nodeName,n=Si.get(t);if(n)return n;Si.set(t,n=[]);for(var r,i=e,a=Element.prototype;a!==i;){for(var o in r=xe(i),r)r[o].set&&n.push(o);i=we(i)}return n}var wi={get(e,t){let n=e.props.length;for(;n--;){let r=e.props[n];if(Ee(r)&&(r=r()),typeof r==`object`&&r&&t in r)return r[t]}},set(e,t,n){let r=e.props.length;for(;r--;){let i=e.props[r];Ee(i)&&(i=i());let a=be(i,t);if(a&&a.set)return a.set(n),!0}return!1},getOwnPropertyDescriptor(e,t){let n=e.props.length;for(;n--;){let r=e.props[n];if(Ee(r)&&(r=r()),typeof r==`object`&&r&&t in r){let e=be(r,t);return e&&!e.configurable&&(e.configurable=!0),e}}},has(e,t){if(t===Ue||t===We)return!1;for(let n of e.props)if(Ee(n)&&(n=n()),n!=null&&t in n)return!0;return!1},ownKeys(e){let t=[];for(let n of e.props)if(Ee(n)&&(n=n()),n){for(let e in n)t.includes(e)||t.push(e);for(let e of Object.getOwnPropertySymbols(n))t.includes(e)||t.push(e)}return t}};function Ti(...e){return new Proxy({props:e},wi)}function Ei(e,t,n,r){var i=!bt||(n&2)!=0,a=(n&8)!=0,o=(n&16)!=0,s=r,c=!0,l=()=>(c&&(c=!1,s=o?Pr(r):r),s);let u;if(a){var d=Ue in e||We in e;u=be(e,t)?.set??(d&&t in e?n=>e[t]=n:void 0)}var f,p=!1;a?[f,p]=Ht(()=>e[t]):f=e[t],f===void 0&&r!==void 0&&(f=l(),u&&(i&&tt(t),u(f)));var m=i?()=>{var n=e[t];return n===void 0?l():(c=!0,n)}:()=>{var n=e[t];return n!==void 0&&(s=void 0),n===void 0?s:n};if(i&&!(n&4))return m;if(u){var h=e.$$legacy;return(function(e,t){return arguments.length>0?((!i||!t||h||p)&&u(t?m():e),e):m()})}var g=!1,_=(n&1?mn:gn)(()=>(g=!1,m()));a&&X(_);var v=G;return(function(e,t){if(arguments.length>0){let n=t?X(_):i&&a?Mn(e):e;return I(_,n),g=!0,s!==void 0&&(s=n),e}return gr&&g||v.f&16384?_.v:X(_)})}typeof window<`u`&&((window.__svelte??={}).v??=new Set).add(`5`);function Di(t){return e.Platform.isMacOS?t.metaKey:t.ctrlKey}var Oi=class{store;plugin;constructor(e,t){this.plugin=t,this.store=zt(0),t.app.workspace.onLayoutReady(()=>{let{vault:n,metadataCache:r,workspace:i}=t.app;e.registerEvent(n.on(`create`,this.bump,this)),e.registerEvent(n.on(`delete`,this.bump,this)),e.registerEvent(n.on(`rename`,this.onRename,this)),e.registerEvent(r.on(`changed`,this.bump,this)),e.registerEvent(i.on(`periodic-notes:resolve`,this.bumpUnconditionally,this)),e.registerEvent(i.on(`periodic-notes:settings-updated`,this.bumpUnconditionally,this)),this.bump()})}bump(e){e&&!this.plugin.isPeriodic(e.path)||this.store.update(e=>e+1)}bumpUnconditionally(){this.store.update(e=>e+1)}onRename(e,t){(this.plugin.isPeriodic(e.path)||this.plugin.isPeriodic(t))&&this.store.update(e=>e+1)}getFile(e,t){return this.plugin.getPeriodicNote(t,e)}isGranularityEnabled(e){return this.plugin.settings.granularities[e]?.enabled??e===`day`}getEnabledGranularities(){return[`week`,`month`,`year`].filter(e=>this.plugin.settings.granularities[e]?.enabled)}};function ki(e,t){return`${e}:${t.format(a[e])}`}function Ai(e,t,n){let r=new Map,i=e[1].days[0];for(let i of e){for(let e of i.days)r.set(ki(`day`,e),t(e,`day`));if(n.includes(`week`)){let e=i.days[0];r.set(ki(`week`,e),t(e,`week`))}}return n.includes(`month`)&&r.set(ki(`month`,i),t(i,`month`)),n.includes(`year`)&&r.set(ki(`year`,i),t(i,`year`)),r}var ji=Q(`<td><div role="button" tabindex="0"> </div></td>`),Mi={hash:`svelte-1l5s8qw`,code:`.day.svelte-1l5s8qw {background-color:var(--color-background-day);border-radius:4px;color:var(--color-text-day);cursor:pointer;font-size:0.8em;height:100%;padding:4px;position:relative;text-align:center;transition:background-color 0.1s ease-in,
+let e=require(`obsidian`);function t(e,t){return`${e}:${t.clone().startOf(e).toISOString()}`}function n(e,t,n){let r=0,i=e.length-1;for(;r<=i;){let a=r+i>>>1,o=e[a];if(o===t)return e[a+(n===`forwards`?1:-1)]??null;o<t?r=a+1:i=a-1}return null}var r=[`day`,`week`,`month`,`year`],i=class{byPath=new Map;byKey=new Map;sortedByGranularity=new Map;dirtyGranularities=new Set(r);set(e){let n=t(e.granularity,e.date),r=this.byPath.get(e.filePath);if(r){let e=t(r.granularity,r.date);e!==n&&(this.byKey.delete(e),this.dirtyGranularities.add(r.granularity))}let i=this.byKey.get(n);i&&i.filePath!==e.filePath&&this.byPath.delete(i.filePath);let a=!this.byKey.has(n);this.byPath.set(e.filePath,e),this.byKey.set(n,e),a&&this.dirtyGranularities.add(e.granularity)}remove(e){let n=this.byPath.get(e);n&&(this.byKey.delete(t(n.granularity,n.date)),this.byPath.delete(e),this.dirtyGranularities.add(n.granularity))}clear(){this.byPath.clear(),this.byKey.clear(),this.sortedByGranularity.clear(),this.dirtyGranularities=new Set(r)}get(e){return e?this.byPath.get(e)??null:null}getByKey(e,n){return this.byKey.get(t(e,n))??null}has(e,t){let n=this.byPath.get(e);return n?t?t===n.granularity:!0:!1}findAdjacent(e,r){let i=this.get(e);if(!i)return null;let a=n(this.getSortedKeys(i.granularity),t(i.granularity,i.date),r);return a?this.byKey.get(a)??null:null}getSortedKeys(e){if(!this.dirtyGranularities.has(e)){let t=this.sortedByGranularity.get(e);if(t)return t}let t=`${e}:`,n=[];for(let e of this.byKey.keys())e.startsWith(t)&&n.push(e);return n.sort(),this.sortedByGranularity.set(e,n),this.dirtyGranularities.delete(e),n}},a={day:`YYYY-MM-DD`,week:`gggg-[W]ww`,month:`YYYY-MM`,year:`YYYY`},o={enabled:!1,format:``,folder:``,templatePath:void 0},s={granularities:{day:{...o},week:{...o},month:{...o},year:{...o}}},c=[`sunday`,`monday`,`tuesday`,`wednesday`,`thursday`,`friday`,`saturday`],l=`calendar`,u=Symbol(`displayedMonth`);function d(e,t){return e.granularities[t].format||a[t]}function f(e,t){let n=e.granularities[t].format;if(!n)return[a[t]];let r=/[^/]*$/.exec(n);return r?[n,r[0]]:[n]}function p(e,t){return e.granularities[t]}function m(e){return r.filter(t=>e.granularities[t].enabled)}function h(e){return e.replace(/\[[^\]]*\]/g,``).replace(/\\./g,``)}function g(e){return e.indexOf(`/`)===-1?e:e.split(`/`).pop()??``}function _(e){return!/[?<>\\:*|"]/g.test(e)&&!/[\x00-\x1f\x80-\x9f]/g.test(e)&&!/^\.+$/.test(e)&&!/^(con|prn|aux|nul|com[0-9]|lpt[0-9])(\..*)?$/i.test(e)}function v(e,t){if(!e)return``;if(!_(e))return`Format contains illegal characters`;if(t===`day`){let t=window.moment().format(e);if(!window.moment(t,e,!0).isValid())return`Failed to parse format`}return``}function y(e){let t=g(e).replace(/\[[^\]]*\]/g,``);return![`M`,`D`].every(e=>t.includes(e))||!(t.includes(`Y`)||t.includes(`y`))}function b(e,t){let n=window.moment().format(e);return window.moment(n,e,!0).isValid()?h(e).includes(`/`)&&t===`day`&&y(e)?`fragile-basename`:`valid`:`loose-parsing`}function ee(...e){let t=[];for(let n=0,r=e.length;n<r;n++)t=t.concat(e[n].split(`/`));let n=[];for(let e=0,r=t.length;e<r;e++){let r=t[e];!r||r===`.`||n.push(r)}return t[0]===``&&n.unshift(``),n.join(`/`)}var x=/{{\s*(date|time)\s*(([-+]\d+)([ymwdhs]))?\s*(:.+?)?}}/gi,te=/{{\s*(month)\s*(([-+]\d+)([ymwdhs]))?\s*(:.+?)?}}/gi,ne=/{{\s*(year)\s*(([-+]\d+)([ymwdhs]))?\s*(:.+?)?}}/gi,re=RegExp(`{{\\s*(${c.join(`|`)})\\s*:(.*?)}}`,`gi`);function ie(){let{moment:e}=window,t=e.localeData().firstDayOfWeek(),n=[...c];for(;t;){let e=n.shift();e&&n.push(e),t--}return n}function ae(e){let t=ie().indexOf(e.toLowerCase());return Math.max(0,t)}function oe(e,t,n,r,i){let a=window.moment();return e.replace(n,(e,n,o,s,c,l)=>{let u=t.clone();return i&&u.startOf(i),u.set({hour:a.get(`hour`),minute:a.get(`minute`),second:a.get(`second`)}),o&&u.add(parseInt(s,10),c),l?u.format(l.substring(1).trim()):u.format(r)})}function se(e,t,n,r,i){let a=i.replace(/{{\s*date\s*}}/gi,e).replace(/{{\s*time\s*}}/gi,window.moment().format(`HH:mm`)).replace(/{{\s*title\s*}}/gi,e);return t===`day`&&(a=a.replace(/{{\s*yesterday\s*}}/gi,n.clone().subtract(1,`day`).format(r)).replace(/{{\s*tomorrow\s*}}/gi,n.clone().add(1,`d`).format(r)),a=oe(a,n,x,r)),t===`week`&&(a=a.replace(re,(e,t,r)=>{let i=ae(t);return n.weekday(i).format(r.trim())})),t===`month`&&(a=oe(a,n,te,r,`month`)),t===`year`&&(a=oe(a,n,ne,r,`year`)),a}async function ce(t,n,r){if(!n||n===`/`)return``;let{metadataCache:i,vault:a}=t,o=(0,e.normalizePath)(n);try{let e=i.getFirstLinkpathDest(o,``);return e?a.cachedRead(e):``}catch(t){return console.error(`[Periodic Notes] Failed to read the ${r} note template '${o}'`,t),new e.Notice(`Failed to read the ${r} note template`),``}}async function le(e,t,n,r){let i=d(n,r.granularity),a=await ce(e,n.granularities[r.granularity].templatePath,r.granularity),o=se(t.basename,r.granularity,r.date,i,a);await e.vault.modify(t,o)}async function ue(t,n,r){let i=(0,e.normalizePath)(ee(r.folder??``,n.endsWith(`.md`)?n:`${n}.md`));return await de(t,i),i}async function de(e,t){let n=t.replace(/\\/g,`/`).split(`/`);n.pop();let r=``;for(let t of n)r=r?`${r}/${t}`:t,e.vault.getAbstractFileByPath(r)||await e.vault.createFolder(r)}function fe(e){let t=e.extension.length+1;return e.path.slice(0,-t)}function pe(e,t,n){if(b(t,n)===`fragile-basename`){let n=fe(e),r=(h(t).match(/\//g)?.length??0)+1;return n.split(`/`).slice(-r).join(`/`)}return e.basename}var me=class extends e.Component{index=new i;constructor(t,n){super(),this.app=t,this.plugin=n,this.app.workspace.onLayoutReady(()=>{console.info(`[Periodic Notes] initializing cache`),this.initialize(),this.registerEvent(this.app.vault.on(`create`,t=>{t instanceof e.TFile&&this.resolve(t,`create`)})),this.registerEvent(this.app.vault.on(`delete`,t=>{t instanceof e.TFile&&this.index.remove(t.path)})),this.registerEvent(this.app.vault.on(`rename`,this.onRename,this)),this.registerEvent(this.app.metadataCache.on(`changed`,this.onMetadataChanged,this)),this.registerEvent(this.app.workspace.on(`periodic-notes:settings-updated`,this.reset,this))})}reset(){console.info(`[Periodic Notes] resetting cache`),this.index.clear(),this.initialize()}initialize(){let t=this.plugin.settings,n=new Set,r=(t,i)=>{if(!n.has(t)){n.add(t);for(let n of t.children)n instanceof e.TFile?i(n):n instanceof e.TFolder&&r(n,i)}},i=m(t);for(let n of i){let i=t.granularities[n].folder||`/`,a=this.app.vault.getAbstractFileByPath(i);a instanceof e.TFolder&&r(a,t=>{if(t instanceof e.TFile){this.resolve(t,`initialize`);let e=this.app.metadataCache.getFileCache(t);e&&this.onMetadataChanged(t,``,e)}})}}onMetadataChanged(t,n,r){let i=this.plugin.settings,a=m(i);if(a.length!==0)for(let n of a){let a=i.granularities[n].folder||`/`;if(!t.path.startsWith(a===`/`?``:`${a}/`))continue;let o=(0,e.parseFrontMatterEntry)(r.frontmatter,n);if(!o)continue;let s=d(i,n);if(typeof o==`string`){let e=window.moment(o,s,!0);e.isValid()&&this.index.set({filePath:t.path,date:e,granularity:n,match:`frontmatter`});return}}}onRename(t,n){t instanceof e.TFile&&(this.index.remove(n),this.resolve(t,`rename`))}resolve(t,n=`create`){let r=this.plugin.settings,i=m(r);if(i.length===0)return;let a=this.index.get(t.path);if(!(a&&a.match===`frontmatter`))for(let a of i){let i=r.granularities[a].folder||`/`;if(!t.path.startsWith(i===`/`?``:`${i}/`))continue;let o=f(r,a),s=pe(t,o[0],a),c=window.moment(s,o,!0);if(c.isValid()){let i={filePath:t.path,date:c,granularity:a,match:`filename`};this.index.set(i),n===`create`&&t.stat.size===0&&le(this.app,t,r,i).catch(n=>{console.error(`[Periodic Notes] failed to apply template`,n),new e.Notice(`Periodic Notes: failed to apply template to "${t.path}". See console for details.`)}),this.app.workspace.trigger(`periodic-notes:resolve`,a,t);return}}}getPeriodicNote(t,n){let r=this.index.getByKey(t,n);if(!r)return null;let i=this.app.vault.getAbstractFileByPath(r.filePath);return i instanceof e.TFile?i:(this.index.remove(r.filePath),null)}isPeriodic(e,t){return this.index.has(e,t)}find(e){return this.index.get(e)}findAdjacent(e,t){return this.index.findAdjacent(e,t)}},he=Array.isArray,ge=Array.prototype.indexOf,_e=Array.prototype.includes,ve=Array.from,ye=Object.defineProperty,be=Object.getOwnPropertyDescriptor,xe=Object.getOwnPropertyDescriptors,Se=Object.prototype,Ce=Array.prototype,we=Object.getPrototypeOf,Te=Object.isExtensible;function Ee(e){return typeof e==`function`}var De=()=>{};function Oe(e){for(var t=0;t<e.length;t++)e[t]()}function ke(){var e,t;return{promise:new Promise((n,r)=>{e=n,t=r}),resolve:e,reject:t}}var S=1024,C=2048,Ae=4096,je=8192,Me=16384,Ne=32768,Pe=1<<25,Fe=65536,Ie=1<<19,Le=1<<20,Re=1<<25,ze=65536,Be=1<<21,Ve=1<<22,He=1<<23,Ue=Symbol(`$state`),We=Symbol(`legacy props`),Ge=Symbol(``),Ke=new class extends Error{name=`StaleReactionError`;message="The reaction that called `getAbortSignal()` was re-run or destroyed"},qe=!!globalThis.document?.contentType&&globalThis.document.contentType.includes(`xml`);function Je(e){throw Error(`https://svelte.dev/e/lifecycle_outside_component`)}function Ye(){throw Error(`https://svelte.dev/e/async_derived_orphan`)}function Xe(e,t,n){throw Error(`https://svelte.dev/e/each_key_duplicate`)}function Ze(e){throw Error(`https://svelte.dev/e/effect_in_teardown`)}function Qe(){throw Error(`https://svelte.dev/e/effect_in_unowned_derived`)}function $e(e){throw Error(`https://svelte.dev/e/effect_orphan`)}function et(){throw Error(`https://svelte.dev/e/effect_update_depth_exceeded`)}function tt(e){throw Error(`https://svelte.dev/e/props_invalid_value`)}function nt(){throw Error(`https://svelte.dev/e/set_context_after_init`)}function rt(){throw Error(`https://svelte.dev/e/state_descriptors_fixed`)}function it(){throw Error(`https://svelte.dev/e/state_prototype_fixed`)}function at(){throw Error(`https://svelte.dev/e/state_unsafe_mutation`)}function ot(){throw Error(`https://svelte.dev/e/svelte_boundary_reset_onerror`)}var st={},w=Symbol(),ct=`http://www.w3.org/1999/xhtml`;function lt(e){console.warn(`https://svelte.dev/e/hydration_mismatch`)}function ut(){console.warn(`https://svelte.dev/e/svelte_boundary_reset_noop`)}var T=!1;function dt(e){T=e}var E;function D(e){if(e===null)throw lt(),st;return E=e}function ft(){return D(In(E))}function O(e){if(T){if(In(E)!==null)throw lt(),st;E=e}}function pt(e=1){if(T){for(var t=e,n=E;t--;)n=In(n);E=n}}function mt(e=!0){for(var t=0,n=E;;){if(n.nodeType===8){var r=n.data;if(r===`]`){if(t===0)return n;--t}else (r===`[`||r===`[!`||r[0]===`[`&&!isNaN(Number(r.slice(1))))&&(t+=1)}var i=In(n);e&&n.remove(),n=i}}function ht(e){if(!e||e.nodeType!==8)throw lt(),st;return e.data}function gt(e){return e===this.v}function _t(e,t){return e==e?e!==t||typeof e==`object`&&!!e||typeof e==`function`:t==t}function vt(e){return!_t(e,this.v)}var yt=!1,bt=!1,k=null;function xt(e){k=e}function St(e){return Dt(`getContext`).get(e)}function Ct(e,t){let n=Dt(`setContext`);if(yt){var r=G.f;!H&&r&32&&!k.i||nt()}return n.set(e,t),t}function wt(e,t=!1,n){k={p:k,i:!1,c:null,e:null,s:e,x:null,r:G,l:bt&&!t?{s:null,u:null,$:[]}:null}}function Tt(e){var t=k,n=t.e;if(n!==null){t.e=null;for(var r of n)Yn(r)}return e!==void 0&&(t.x=e),t.i=!0,k=t.p,e??{}}function Et(){return!bt||k!==null&&k.l===null}function Dt(e){return k===null&&Je(e),k.c??=new Map(Ot(k)||void 0)}function Ot(e){let t=e.p;for(;t!==null;){let e=t.c;if(e!==null)return e;t=t.p}return null}var kt=[];function At(){var e=kt;kt=[],Oe(e)}function jt(e){if(kt.length===0&&!Ht){var t=kt;queueMicrotask(()=>{t===kt&&At()})}kt.push(e)}function Mt(e){var t=G;if(t===null)return H.f|=He,e;if(!(t.f&32768)&&!(t.f&4))throw e;Nt(e,t)}function Nt(e,t){for(;t!==null;){if(t.f&128){if(!(t.f&32768))throw e;try{t.b.error(e);return}catch(t){e=t}}t=t.parent}throw e}var Pt=~(C|Ae|S);function A(e,t){e.f=e.f&Pt|t}function Ft(e){e.f&512||e.deps===null?A(e,S):A(e,Ae)}function It(e){if(e!==null)for(let t of e)!(t.f&2)||!(t.f&65536)||(t.f^=ze,It(t.deps))}function Lt(e,t,n){e.f&2048?t.add(e):e.f&4096&&n.add(e),It(e.deps),A(e,S)}var Rt=!1,zt=!1;function Bt(e){var t=zt;try{return zt=!1,[e(),zt]}finally{zt=t}}var j=new Set,M=null,N=null,Vt=null,Ht=!1,Ut=!1,Wt=null,Gt=null,Kt=0,qt=1,Jt=class e{id=qt++;current=new Map;previous=new Map;#e=new Set;#t=new Set;#n=new Map;#r=new Map;#i=null;#a=[];#o=[];#s=new Set;#c=new Set;#l=new Map;is_fork=!1;#u=!1;#d=new Set;#f(){return this.is_fork||this.#r.size>0}#p(){for(let n of this.#d)for(let r of n.#r.keys()){for(var e=!1,t=r;t.parent!==null;){if(this.#l.has(t)){e=!0;break}t=t.parent}if(!e)return!0}return!1}skip_effect(e){this.#l.has(e)||this.#l.set(e,{d:[],m:[]})}unskip_effect(e){var t=this.#l.get(e);if(t){this.#l.delete(e);for(var n of t.d)A(n,C),this.schedule(n);for(n of t.m)A(n,Ae),this.schedule(n)}}#m(){if(Kt++>1e3&&(j.delete(this),Yt()),!this.#f()){for(let e of this.#s)this.#c.delete(e),A(e,C),this.schedule(e);for(let e of this.#c)A(e,Ae),this.schedule(e)}let t=this.#a;this.#a=[],this.apply();var n=Wt=[],r=[],i=Gt=[];for(let e of t)try{this.#h(e,n,r)}catch(t){throw tn(e),t}if(M=null,i.length>0){var a=e.ensure();for(let e of i)a.schedule(e)}if(Wt=null,Gt=null,this.#f()||this.#p()){this.#g(r),this.#g(n);for(let[e,t]of this.#l)en(e,t)}else{this.#n.size===0&&j.delete(this),this.#s.clear(),this.#c.clear();for(let e of this.#e)e(this);this.#e.clear(),Xt(r),Xt(n),this.#i?.resolve()}var o=M;if(this.#a.length>0){let e=o??=this;e.#a.push(...this.#a.filter(t=>!e.#a.includes(t)))}o!==null&&(j.add(o),o.#m()),j.has(this)||this.#_()}#h(e,t,n){e.f^=S;for(var r=e.first;r!==null;){var i=r.f,a=(i&96)!=0;if(!(a&&i&1024||i&8192||this.#l.has(r))&&r.fn!==null){a?r.f^=S:i&4?t.push(r):yt&&i&16777224?n.push(r):wr(r)&&(i&16&&this.#c.add(r),kr(r));var o=r.first;if(o!==null){r=o;continue}}for(;r!==null;){var s=r.next;if(s!==null){r=s;break}r=r.parent}}}#g(e){for(var t=0;t<e.length;t+=1)Lt(e[t],this.#s,this.#c)}capture(e,t,n=!1){t!==w&&!this.previous.has(e)&&this.previous.set(e,t),e.f&8388608||(this.current.set(e,[e.v,n]),N?.set(e,e.v))}activate(){M=this}deactivate(){M=null,N=null}flush(){try{Ut=!0,M=this,this.#m()}finally{Kt=0,Vt=null,Wt=null,Gt=null,Ut=!1,M=null,N=null,xn.clear()}}discard(){for(let e of this.#t)e(this);this.#t.clear(),j.delete(this)}register_created_effect(e){this.#o.push(e)}#_(){for(let l of j){var e=l.id<this.id,t=[];for(let[r,[i,a]]of this.current){if(l.current.has(r)){var n=l.current.get(r)[0];if(e&&i!==n)l.current.set(r,[i,a]);else continue}t.push(r)}var r=[...l.current.keys()].filter(e=>!this.current.has(e));if(r.length===0)e&&l.discard();else if(t.length>0){l.activate();var i=new Set,a=new Map;for(var o of t)Zt(o,r,i,a);a=new Map;var s=[...l.current.keys()].filter(e=>this.current.has(e)?this.current.get(e)[0]!==e:!0);for(let e of this.#o)!(e.f&155648)&&Qt(e,s,a)&&(e.f&4194320?(A(e,C),l.schedule(e)):l.#s.add(e));if(l.#a.length>0){l.apply();for(var c of l.#a)l.#h(c,[],[]);l.#a=[]}l.deactivate()}}for(let e of j)e.#d.has(this)&&(e.#d.delete(this),e.#d.size===0&&!e.#f()&&(e.activate(),e.#m()))}increment(e,t){let n=this.#n.get(t)??0;if(this.#n.set(t,n+1),e){let e=this.#r.get(t)??0;this.#r.set(t,e+1)}}decrement(e,t,n){let r=this.#n.get(t)??0;if(r===1?this.#n.delete(t):this.#n.set(t,r-1),e){let e=this.#r.get(t)??0;e===1?this.#r.delete(t):this.#r.set(t,e-1)}this.#u||n||(this.#u=!0,jt(()=>{this.#u=!1,this.flush()}))}transfer_effects(e,t){for(let t of e)this.#s.add(t);for(let e of t)this.#c.add(e);e.clear(),t.clear()}oncommit(e){this.#e.add(e)}ondiscard(e){this.#t.add(e)}settled(){return(this.#i??=ke()).promise}static ensure(){if(M===null){let t=M=new e;Ut||(j.add(M),Ht||jt(()=>{M===t&&t.flush()}))}return M}apply(){if(!yt||!this.is_fork&&j.size===1){N=null;return}N=new Map;for(let[e,[t]]of this.current)N.set(e,t);for(let n of j)if(!(n===this||n.is_fork)){var e=!1,t=!1;if(n.id<this.id)for(let[r,[,i]]of n.current)i||(e||=this.current.has(r),t||=!this.current.has(r));if(e&&t)this.#d.add(n);else for(let[e,t]of n.previous)N.has(e)||N.set(e,t)}}schedule(e){if(Vt=e,e.b?.is_pending&&e.f&16777228&&!(e.f&32768)){e.b.defer_effect(e);return}for(var t=e;t.parent!==null;){t=t.parent;var n=t.f;if(Wt!==null&&t===G&&(yt||(H===null||!(H.f&2))&&!Rt))return;if(n&96){if(!(n&1024))return;t.f^=S}}this.#a.push(t)}};function Yt(){try{et()}catch(e){Nt(e,Vt)}}var P=null;function Xt(e){var t=e.length;if(t!==0){for(var n=0;n<t;){var r=e[n++];if(!(r.f&24576)&&wr(r)&&(P=new Set,kr(r),r.deps===null&&r.first===null&&r.nodes===null&&r.teardown===null&&r.ac===null&&or(r),P?.size>0)){xn.clear();for(let e of P){if(e.f&24576)continue;let t=[e],n=e.parent;for(;n!==null;)P.has(n)&&(P.delete(n),t.push(n)),n=n.parent;for(let e=t.length-1;e>=0;e--){let n=t[e];n.f&24576||kr(n)}}P.clear()}}P=null}}function Zt(e,t,n,r){if(!n.has(e)&&(n.add(e),e.reactions!==null))for(let i of e.reactions){let e=i.f;e&2?Zt(i,t,n,r):e&4194320&&!(e&2048)&&Qt(i,t,r)&&(A(i,C),$t(i))}}function Qt(e,t,n){let r=n.get(e);if(r!==void 0)return r;if(e.deps!==null)for(let r of e.deps){if(_e.call(t,r))return!0;if(r.f&2&&Qt(r,t,n))return n.set(r,!0),!0}return n.set(e,!1),!1}function $t(e){M.schedule(e)}function en(e,t){if(!(e.f&32&&e.f&1024)){e.f&2048?t.d.push(e):e.f&4096&&t.m.push(e),A(e,S);for(var n=e.first;n!==null;)en(n,t),n=n.next}}function tn(e){A(e,S);for(var t=e.first;t!==null;)tn(t),t=t.next}function nn(e){let t=0,n=Cn(0),r;return()=>{Kn()&&(X(n),$n(()=>(t===0&&(r=Mr(()=>e(()=>Dn(n)))),t+=1,()=>{jt(()=>{--t,t===0&&(r?.(),r=void 0,Dn(n))})})))}}var rn=Fe|Ie;function an(e,t,n,r){new on(e,t,n,r)}var on=class{parent;is_pending=!1;transform_error;#e;#t=T?E:null;#n;#r;#i;#a=null;#o=null;#s=null;#c=null;#l=0;#u=0;#d=!1;#f=new Set;#p=new Set;#m=null;#h=nn(()=>(this.#m=Cn(this.#l),()=>{this.#m=null}));constructor(e,t,n,r){this.#e=e,this.#n=t,this.#r=e=>{var t=G;t.b=this,t.f|=128,n(e)},this.parent=G.b,this.transform_error=r??this.parent?.transform_error??(e=>e),this.#i=tr(()=>{if(T){let e=this.#t;ft();let t=e.data===`[!`;if(e.data.startsWith(`[?`)){let t=JSON.parse(e.data.slice(2));this.#_(t)}else t?this.#v():this.#g()}else this.#y()},rn),T&&(this.#e=E)}#g(){try{this.#a=B(()=>this.#r(this.#e))}catch(e){this.error(e)}}#_(e){let t=this.#n.failed;t&&(this.#s=B(()=>{t(this.#e,()=>e,()=>()=>{})}))}#v(){let e=this.#n.pending;e&&(this.is_pending=!0,this.#o=B(()=>e(this.#e)),jt(()=>{var e=this.#c=document.createDocumentFragment(),t=R();e.append(t),this.#a=this.#x(()=>B(()=>this.#r(t))),this.#u===0&&(this.#e.before(e),this.#c=null,sr(this.#o,()=>{this.#o=null}),this.#b(M))}))}#y(){try{if(this.is_pending=this.has_pending_snippet(),this.#u=0,this.#l=0,this.#a=B(()=>{this.#r(this.#e)}),this.#u>0){var e=this.#c=document.createDocumentFragment();dr(this.#a,e);let t=this.#n.pending;this.#o=B(()=>t(this.#e))}else this.#b(M)}catch(e){this.error(e)}}#b(e){this.is_pending=!1,e.transfer_effects(this.#f,this.#p)}defer_effect(e){Lt(e,this.#f,this.#p)}is_rendered(){return!this.is_pending&&(!this.parent||this.parent.is_rendered())}has_pending_snippet(){return!!this.#n.pending}#x(e){var t=G,n=H,r=k;gr(this.#i),W(this.#i),xt(this.#i.ctx);try{return Jt.ensure(),e()}catch(e){return Mt(e),null}finally{gr(t),W(n),xt(r)}}#S(e,t){if(!this.has_pending_snippet()){this.parent&&this.parent.#S(e,t);return}this.#u+=e,this.#u===0&&(this.#b(t),this.#o&&sr(this.#o,()=>{this.#o=null}),this.#c&&=(this.#e.before(this.#c),null))}update_pending_count(e,t){this.#S(e,t),this.#l+=e,!(!this.#m||this.#d)&&(this.#d=!0,jt(()=>{this.#d=!1,this.#m&&Tn(this.#m,this.#l)}))}get_effect_pending(){return this.#h(),X(this.#m)}error(e){var t=this.#n.onerror;let n=this.#n.failed;if(!t&&!n)throw e;this.#a&&=(V(this.#a),null),this.#o&&=(V(this.#o),null),this.#s&&=(V(this.#s),null),T&&(D(this.#t),pt(),D(mt()));var r=!1,i=!1;let a=()=>{if(r){ut();return}r=!0,i&&ot(),this.#s!==null&&sr(this.#s,()=>{this.#s=null}),this.#x(()=>{this.#y()})},o=e=>{try{i=!0,t?.(e,a),i=!1}catch(e){Nt(e,this.#i&&this.#i.parent)}n&&(this.#s=this.#x(()=>{try{return B(()=>{var t=G;t.b=this,t.f|=128,n(this.#e,()=>e,()=>a)})}catch(e){return Nt(e,this.#i.parent),null}}))};jt(()=>{var t;try{t=this.transform_error(e)}catch(e){Nt(e,this.#i&&this.#i.parent);return}typeof t==`object`&&t&&typeof t.then==`function`?t.then(o,e=>Nt(e,this.#i&&this.#i.parent)):o(t)})}};function sn(e,t,n,r){let i=Et()?dn:pn;var a=e.filter(e=>!e.settled);if(n.length===0&&a.length===0){r(t.map(i));return}var o=G,s=cn(),c=a.length===1?a[0].promise:a.length>1?Promise.all(a.map(e=>e.promise)):null;function l(e){s();try{r(e)}catch(e){o.f&16384||Nt(e,o)}ln()}if(n.length===0){c.then(()=>l(t.map(i)));return}var u=un();function d(){Promise.all(n.map(e=>fn(e))).then(e=>l([...t.map(i),...e])).catch(e=>Nt(e,o)).finally(()=>u())}c?c.then(()=>{s(),d(),ln()}):d()}function cn(){var e=G,t=H,n=k,r=M;return function(i=!0){gr(e),W(t),xt(n),i&&!(e.f&16384)&&(r?.activate(),r?.apply())}}function ln(e=!0){gr(null),W(null),xt(null),e&&M?.deactivate()}function un(){var e=G,t=e.b,n=M,r=t.is_rendered();return t.update_pending_count(1,n),n.increment(r,e),(i=!1)=>{t.update_pending_count(-1,n),n.decrement(r,e,i)}}function dn(e){var t=2|C,n=H!==null&&H.f&2?H:null;return G!==null&&(G.f|=Ie),{ctx:k,deps:null,effects:null,equals:gt,f:t,fn:e,reactions:null,rv:0,v:w,wv:0,parent:n??G,ac:null}}function fn(e,t,n){let r=G;r===null&&Ye();var i=void 0,a=Cn(w),o=!H,s=new Map;return Qn(()=>{var t=G,n=ke();i=n.promise;try{Promise.resolve(e()).then(n.resolve,n.reject).finally(ln)}catch(e){n.reject(e),ln()}var c=M;if(o){if(t.f&32768)var l=un();if(r.b.is_rendered())s.get(c)?.reject(Ke),s.delete(c);else{for(let e of s.values())e.reject(Ke);s.clear()}s.set(c,n)}let u=(e,n=void 0)=>{if(l&&l(n===Ke),!(n===Ke||t.f&16384)){if(c.activate(),n)a.f|=He,Tn(a,n);else{a.f&8388608&&(a.f^=He),Tn(a,e);for(let[e,t]of s){if(s.delete(e),e===c)break;t.reject(Ke)}}c.deactivate()}};n.promise.then(u,e=>u(null,e||`unknown`))}),qn(()=>{for(let e of s.values())e.reject(Ke)}),new Promise(e=>{function t(n){function r(){n===i?e(a):t(i)}n.then(r,r)}t(i)})}function F(e){let t=dn(e);return yt||_r(t),t}function pn(e){let t=dn(e);return t.equals=vt,t}function mn(e){var t=e.effects;if(t!==null){e.effects=null;for(var n=0;n<t.length;n+=1)V(t[n])}}function hn(e){for(var t=e.parent;t!==null;){if(!(t.f&2))return t.f&16384?null:t;t=t.parent}return null}function gn(e){var t,n=G;gr(hn(e));try{e.f&=~ze,mn(e),t=Er(e)}finally{gr(n)}return t}function _n(e){var t=e.v,n=gn(e);if(!e.equals(n)&&(e.wv=Cr(),(!M?.is_fork||e.deps===null)&&(e.v=n,M?.capture(e,t,!0),e.deps===null))){A(e,S);return}mr||(N===null?Ft(e):(Kn()||M?.is_fork)&&N.set(e,n))}function vn(e){if(e.effects!==null)for(let t of e.effects)(t.teardown||t.ac)&&(t.teardown?.(),t.ac?.abort(Ke),t.teardown=De,t.ac=null,Or(t,0),rr(t))}function yn(e){if(e.effects!==null)for(let t of e.effects)t.teardown&&kr(t)}var bn=new Set,xn=new Map,Sn=!1;function Cn(e,t){return{f:0,v:e,reactions:null,equals:gt,rv:0,wv:0}}function I(e,t){let n=Cn(e,t);return _r(n),n}function wn(e,t=!1,n=!0){let r=Cn(e);return t||(r.equals=vt),bt&&n&&k!==null&&k.l!==null&&(k.l.s??=[]).push(r),r}function L(e,t,n=!1){return H!==null&&(!U||H.f&131072)&&Et()&&H.f&4325394&&(K===null||!_e.call(K,e))&&at(),Tn(e,n?kn(t):t,Gt)}function Tn(e,t,n=null){if(!e.equals(t)){var r=e.v;mr?xn.set(e,t):xn.set(e,r),e.v=t;var i=Jt.ensure();if(i.capture(e,r),e.f&2){let t=e;e.f&2048&&gn(t),N===null&&Ft(t)}e.wv=Cr(),On(e,C,n),Et()&&G!==null&&G.f&1024&&!(G.f&96)&&(Y===null?vr([e]):Y.push(e)),!i.is_fork&&bn.size>0&&!Sn&&En()}return t}function En(){Sn=!1;for(let e of bn)e.f&1024&&A(e,Ae),wr(e)&&kr(e);bn.clear()}function Dn(e){L(e,e.v+1)}function On(e,t,n){var r=e.reactions;if(r!==null)for(var i=Et(),a=r.length,o=0;o<a;o++){var s=r[o],c=s.f;if(!(!i&&s===G)){var l=(c&C)===0;if(l&&A(s,t),c&2){var u=s;N?.delete(u),c&65536||(c&512&&(s.f|=ze),On(u,Ae,n))}else if(l){var d=s;c&16&&P!==null&&P.add(d),n===null?$t(d):n.push(d)}}}}function kn(e){if(typeof e!=`object`||!e||Ue in e)return e;let t=we(e);if(t!==Se&&t!==Ce)return e;var n=new Map,r=he(e),i=I(0),a=null,o=xr,s=e=>{if(xr===o)return e();var t=H,n=xr;W(null),Sr(o);var r=e();return W(t),Sr(n),r};return r&&n.set(`length`,I(e.length,a)),new Proxy(e,{defineProperty(e,t,r){(!(`value`in r)||r.configurable===!1||r.enumerable===!1||r.writable===!1)&&rt();var i=n.get(t);return i===void 0?s(()=>{var e=I(r.value,a);return n.set(t,e),e}):L(i,r.value,!0),!0},deleteProperty(e,t){var r=n.get(t);if(r===void 0){if(t in e){let e=s(()=>I(w,a));n.set(t,e),Dn(i)}}else L(r,w),Dn(i);return!0},get(t,r,i){if(r===Ue)return e;var o=n.get(r),c=r in t;if(o===void 0&&(!c||be(t,r)?.writable)&&(o=s(()=>I(kn(c?t[r]:w),a)),n.set(r,o)),o!==void 0){var l=X(o);return l===w?void 0:l}return Reflect.get(t,r,i)},getOwnPropertyDescriptor(e,t){var r=Reflect.getOwnPropertyDescriptor(e,t);if(r&&`value`in r){var i=n.get(t);i&&(r.value=X(i))}else if(r===void 0){var a=n.get(t),o=a?.v;if(a!==void 0&&o!==w)return{enumerable:!0,configurable:!0,value:o,writable:!0}}return r},has(e,t){if(t===Ue)return!0;var r=n.get(t),i=r!==void 0&&r.v!==w||Reflect.has(e,t);return(r!==void 0||G!==null&&(!i||be(e,t)?.writable))&&(r===void 0&&(r=s(()=>I(i?kn(e[t]):w,a)),n.set(t,r)),X(r)===w)?!1:i},set(e,t,o,c){var l=n.get(t),u=t in e;if(r&&t===`length`)for(var d=o;d<l.v;d+=1){var f=n.get(d+``);f===void 0?d in e&&(f=s(()=>I(w,a)),n.set(d+``,f)):L(f,w)}if(l===void 0)(!u||be(e,t)?.writable)&&(l=s(()=>I(void 0,a)),L(l,kn(o)),n.set(t,l));else{u=l.v!==w;var p=s(()=>kn(o));L(l,p)}var m=Reflect.getOwnPropertyDescriptor(e,t);if(m?.set&&m.set.call(c,o),!u){if(r&&typeof t==`string`){var h=n.get(`length`),g=Number(t);Number.isInteger(g)&&g>=h.v&&L(h,g+1)}Dn(i)}return!0},ownKeys(e){X(i);var t=Reflect.ownKeys(e).filter(e=>{var t=n.get(e);return t===void 0||t.v!==w});for(var[r,a]of n)a.v!==w&&!(r in e)&&t.push(r);return t},setPrototypeOf(){it()}})}var An,jn,Mn,Nn;function Pn(){if(An===void 0){An=window,jn=/Firefox/.test(navigator.userAgent);var e=Element.prototype,t=Node.prototype,n=Text.prototype;Mn=be(t,`firstChild`).get,Nn=be(t,`nextSibling`).get,Te(e)&&(e.__click=void 0,e.__className=void 0,e.__attributes=null,e.__style=void 0,e.__e=void 0),Te(n)&&(n.__t=void 0)}}function R(e=``){return document.createTextNode(e)}function Fn(e){return Mn.call(e)}function In(e){return Nn.call(e)}function z(e,t){if(!T)return Fn(e);var n=Fn(E);if(n===null)n=E.appendChild(R());else if(t&&n.nodeType!==3){var r=R();return n?.before(r),D(r),r}return t&&Vn(n),D(n),n}function Ln(e,t=1,n=!1){let r=T?E:e;for(var i;t--;)i=r,r=In(r);if(!T)return r;if(n){if(r?.nodeType!==3){var a=R();return r===null?i?.after(a):r.before(a),D(a),a}Vn(r)}return D(r),r}function Rn(e){e.textContent=``}function zn(){return!yt||P!==null?!1:(G.f&Ne)!==0}function Bn(e,t,n){let r=n?{is:n}:void 0;return document.createElementNS(t??`http://www.w3.org/1999/xhtml`,e,r)}function Vn(e){if(e.nodeValue.length<65536)return;let t=e.nextSibling;for(;t!==null&&t.nodeType===3;)t.remove(),e.nodeValue+=t.nodeValue,t=e.nextSibling}function Hn(e){var t=H,n=G;W(null),gr(null);try{return e()}finally{W(t),gr(n)}}function Un(e){G===null&&(H===null&&$e(e),Qe()),mr&&Ze(e)}function Wn(e,t){var n=t.last;n===null?t.last=t.first=e:(n.next=e,e.prev=n,t.last=e)}function Gn(e,t){var n=G;n!==null&&n.f&8192&&(e|=je);var r={ctx:k,deps:null,nodes:null,f:e|C|512,first:null,fn:t,last:null,next:null,parent:n,b:n&&n.b,prev:null,teardown:null,wv:0,ac:null};M?.register_created_effect(r);var i=r;if(e&4)Wt===null?Jt.ensure().schedule(r):Wt.push(r);else if(t!==null){try{kr(r)}catch(e){throw V(r),e}i.deps===null&&i.teardown===null&&i.nodes===null&&i.first===i.last&&!(i.f&524288)&&(i=i.first,e&16&&e&65536&&i!==null&&(i.f|=Fe))}if(i!==null&&(i.parent=n,n!==null&&Wn(i,n),H!==null&&H.f&2&&!(e&64))){var a=H;(a.effects??=[]).push(i)}return r}function Kn(){return H!==null&&!U}function qn(e){let t=Gn(8,null);return A(t,S),t.teardown=e,t}function Jn(e){Un(`$effect`);var t=G.f;if(!H&&t&32&&!(t&32768)){var n=k;(n.e??=[]).push(e)}else return Yn(e)}function Yn(e){return Gn(4|Le,e)}function Xn(e){Jt.ensure();let t=Gn(64|Ie,e);return(e={})=>new Promise(n=>{e.outro?sr(t,()=>{V(t),n(void 0)}):(V(t),n(void 0))})}function Zn(e){return Gn(4,e)}function Qn(e){return Gn(Ve|Ie,e)}function $n(e,t=0){return Gn(8|t,e)}function er(e,t=[],n=[],r=[]){sn(r,t,n,t=>{Gn(8,()=>e(...t.map(X)))})}function tr(e,t=0){return Gn(16|t,e)}function B(e){return Gn(32|Ie,e)}function nr(e){var t=e.teardown;if(t!==null){let e=mr,n=H;hr(!0),W(null);try{t.call(null)}finally{hr(e),W(n)}}}function rr(e,t=!1){var n=e.first;for(e.first=e.last=null;n!==null;){let e=n.ac;e!==null&&Hn(()=>{e.abort(Ke)});var r=n.next;n.f&64?n.parent=null:V(n,t),n=r}}function ir(e){for(var t=e.first;t!==null;){var n=t.next;t.f&32||V(t),t=n}}function V(e,t=!0){var n=!1;(t||e.f&262144)&&e.nodes!==null&&e.nodes.end!==null&&(ar(e.nodes.start,e.nodes.end),n=!0),A(e,Pe),rr(e,t&&!n),Or(e,0);var r=e.nodes&&e.nodes.t;if(r!==null)for(let e of r)e.stop();nr(e),e.f^=Pe,e.f|=Me;var i=e.parent;i!==null&&i.first!==null&&or(e),e.next=e.prev=e.teardown=e.ctx=e.deps=e.fn=e.nodes=e.ac=e.b=null}function ar(e,t){for(;e!==null;){var n=e===t?null:In(e);e.remove(),e=n}}function or(e){var t=e.parent,n=e.prev,r=e.next;n!==null&&(n.next=r),r!==null&&(r.prev=n),t!==null&&(t.first===e&&(t.first=r),t.last===e&&(t.last=n))}function sr(e,t,n=!0){var r=[];cr(e,r,!0);var i=()=>{n&&V(e),t&&t()},a=r.length;if(a>0){var o=()=>--a||i();for(var s of r)s.out(o)}else i()}function cr(e,t,n){if(!(e.f&8192)){e.f^=je;var r=e.nodes&&e.nodes.t;if(r!==null)for(let e of r)(e.is_global||n)&&t.push(e);for(var i=e.first;i!==null;){var a=i.next,o=(i.f&65536)!=0||(i.f&32)!=0&&(e.f&16)!=0;cr(i,t,o?n:!1),i=a}}}function lr(e){ur(e,!0)}function ur(e,t){if(e.f&8192){e.f^=je,e.f&1024||(A(e,C),Jt.ensure().schedule(e));for(var n=e.first;n!==null;){var r=n.next,i=(n.f&65536)!=0||(n.f&32)!=0;ur(n,i?t:!1),n=r}var a=e.nodes&&e.nodes.t;if(a!==null)for(let e of a)(e.is_global||t)&&e.in()}}function dr(e,t){if(e.nodes)for(var n=e.nodes.start,r=e.nodes.end;n!==null;){var i=n===r?null:In(n);t.append(n),n=i}}var fr=null,pr=!1,mr=!1;function hr(e){mr=e}var H=null,U=!1;function W(e){H=e}var G=null;function gr(e){G=e}var K=null;function _r(e){H!==null&&(!yt||H.f&2)&&(K===null?K=[e]:K.push(e))}var q=null,J=0,Y=null;function vr(e){Y=e}var yr=1,br=0,xr=br;function Sr(e){xr=e}function Cr(){return++yr}function wr(e){var t=e.f;if(t&2048)return!0;if(t&2&&(e.f&=~ze),t&4096){for(var n=e.deps,r=n.length,i=0;i<r;i++){var a=n[i];if(wr(a)&&_n(a),a.wv>e.wv)return!0}t&512&&N===null&&A(e,S)}return!1}function Tr(e,t,n=!0){var r=e.reactions;if(r!==null&&!(!yt&&K!==null&&_e.call(K,e)))for(var i=0;i<r.length;i++){var a=r[i];a.f&2?Tr(a,t,!1):t===a&&(n?A(a,C):a.f&1024&&A(a,Ae),$t(a))}}function Er(e){var t=q,n=J,r=Y,i=H,a=K,o=k,s=U,c=xr,l=e.f;q=null,J=0,Y=null,H=l&96?null:e,K=null,xt(e.ctx),U=!1,xr=++br,e.ac!==null&&(Hn(()=>{e.ac.abort(Ke)}),e.ac=null);try{e.f|=Be;var u=e.fn,d=u();e.f|=Ne;var f=e.deps,p=M?.is_fork;if(q!==null){var m;if(p||Or(e,J),f!==null&&J>0)for(f.length=J+q.length,m=0;m<q.length;m++)f[J+m]=q[m];else e.deps=f=q;if(Kn()&&e.f&512)for(m=J;m<f.length;m++)(f[m].reactions??=[]).push(e)}else !p&&f!==null&&J<f.length&&(Or(e,J),f.length=J);if(Et()&&Y!==null&&!U&&f!==null&&!(e.f&6146))for(m=0;m<Y.length;m++)Tr(Y[m],e);if(i!==null&&i!==e){if(br++,i.deps!==null)for(let e=0;e<n;e+=1)i.deps[e].rv=br;if(t!==null)for(let e of t)e.rv=br;Y!==null&&(r===null?r=Y:r.push(...Y))}return e.f&8388608&&(e.f^=He),d}catch(e){return Mt(e)}finally{e.f^=Be,q=t,J=n,Y=r,H=i,K=a,xt(o),U=s,xr=c}}function Dr(e,t){let n=t.reactions;if(n!==null){var r=ge.call(n,e);if(r!==-1){var i=n.length-1;i===0?n=t.reactions=null:(n[r]=n[i],n.pop())}}if(n===null&&t.f&2&&(q===null||!_e.call(q,t))){var a=t;a.f&512&&(a.f^=512,a.f&=~ze),Ft(a),vn(a),Or(a,0)}}function Or(e,t){var n=e.deps;if(n!==null)for(var r=t;r<n.length;r++)Dr(e,n[r])}function kr(e){var t=e.f;if(!(t&16384)){A(e,S);var n=G,r=pr;G=e,pr=!0;try{t&16777232?ir(e):rr(e),nr(e);var i=Er(e);e.teardown=typeof i==`function`?i:null,e.wv=yr}finally{pr=r,G=n}}}function X(e){var t=(e.f&2)!=0;if(fr?.add(e),H!==null&&!U&&!(G!==null&&G.f&16384)&&(K===null||!_e.call(K,e))){var n=H.deps;if(H.f&2097152)e.rv<br&&(e.rv=br,q===null&&n!==null&&n[J]===e?J++:q===null?q=[e]:q.push(e));else{(H.deps??=[]).push(e);var r=e.reactions;r===null?e.reactions=[H]:_e.call(r,H)||r.push(H)}}if(mr&&xn.has(e))return xn.get(e);if(t){var i=e;if(mr){var a=i.v;return(!(i.f&1024)&&i.reactions!==null||jr(i))&&(a=gn(i)),xn.set(i,a),a}var o=(i.f&512)==0&&!U&&H!==null&&(pr||(H.f&512)!=0),s=(i.f&Ne)===0;wr(i)&&(o&&(i.f|=512),_n(i)),o&&!s&&(yn(i),Ar(i))}if(N?.has(e))return N.get(e);if(e.f&8388608)throw e.v;return e.v}function Ar(e){if(e.f|=512,e.deps!==null)for(let t of e.deps)(t.reactions??=[]).push(e),t.f&2&&!(t.f&512)&&(yn(t),Ar(t))}function jr(e){if(e.v===w)return!0;if(e.deps===null)return!1;for(let t of e.deps)if(xn.has(t)||t.f&2&&jr(t))return!0;return!1}function Mr(e){var t=U;try{return U=!0,e()}finally{U=t}}[...`allowfullscreen.async.autofocus.autoplay.checked.controls.default.disabled.formnovalidate.indeterminate.inert.ismap.loop.multiple.muted.nomodule.novalidate.open.playsinline.readonly.required.reversed.seamless.selected.webkitdirectory.defer.disablepictureinpicture.disableremoteplayback`.split(`.`)];var Nr=[`touchstart`,`touchmove`];function Pr(e){return Nr.includes(e)}var Fr=Symbol(`events`),Ir=new Set,Lr=new Set;function Rr(e,t,n,r={}){function i(e){if(r.capture||Hr.call(t,e),!e.cancelBubble)return Hn(()=>n?.call(this,e))}return e.startsWith(`pointer`)||e.startsWith(`touch`)||e===`wheel`?jt(()=>{t.addEventListener(e,i,r)}):t.addEventListener(e,i,r),i}function zr(e,t,n,r,i){var a={capture:r,passive:i},o=Rr(e,t,n,a);(t===document.body||t===window||t===document||t instanceof HTMLMediaElement)&&qn(()=>{t.removeEventListener(e,o,a)})}function Z(e,t,n){(t[Fr]??={})[e]=n}function Br(e){for(var t=0;t<e.length;t++)Ir.add(e[t]);for(var n of Lr)n(e)}var Vr=null;function Hr(e){var t=this,n=t.ownerDocument,r=e.type,i=e.composedPath?.()||[],a=i[0]||e.target;Vr=e;var o=0,s=Vr===e&&e[Fr];if(s){var c=i.indexOf(s);if(c!==-1&&(t===document||t===window)){e[Fr]=t;return}var l=i.indexOf(t);if(l===-1)return;c<=l&&(o=c)}if(a=i[o]||e.target,a!==t){ye(e,`currentTarget`,{configurable:!0,get(){return a||n}});var u=H,d=G;W(null),gr(null);try{for(var f,p=[];a!==null;){var m=a.assignedSlot||a.parentNode||a.host||null;try{var h=a[Fr]?.[r];h!=null&&(!a.disabled||e.target===a)&&h.call(a,e)}catch(e){f?p.push(e):f=e}if(e.cancelBubble||m===t||m===null)break;a=m}if(f){for(let e of p)queueMicrotask(()=>{throw e});throw f}}finally{e[Fr]=t,delete e.currentTarget,W(u),gr(d)}}}var Ur=globalThis?.window?.trustedTypes&&globalThis.window.trustedTypes.createPolicy(`svelte-trusted-html`,{createHTML:e=>e});function Wr(e){return Ur?.createHTML(e)??e}function Gr(e){var t=Bn(`template`);return t.innerHTML=Wr(e.replaceAll(`<!>`,`<!---->`)),t.content}function Kr(e,t){var n=G;n.nodes===null&&(n.nodes={start:e,end:t,a:null,t:null})}function Q(e,t){var n=(t&1)!=0,r=(t&2)!=0,i,a=!e.startsWith(`<!>`);return()=>{if(T)return Kr(E,null),E;i===void 0&&(i=Gr(a?e:`<!>`+e),n||(i=Fn(i)));var t=r||jn?document.importNode(i,!0):i.cloneNode(!0);if(n){var o=Fn(t),s=t.lastChild;Kr(o,s)}else Kr(t,t);return t}}function $(e,t){if(T){var n=G;(!(n.f&32768)||n.nodes.end===null)&&(n.nodes.end=E),ft();return}e!==null&&e.before(t)}function qr(e,t){var n=t==null?``:typeof t==`object`?`${t}`:t;n!==(e.__t??=e.nodeValue)&&(e.__t=n,e.nodeValue=`${n}`)}function Jr(e,t){return Xr(e,t)}var Yr=new Map;function Xr(e,{target:t,anchor:n,props:r={},events:i,context:a,intro:o=!0,transformError:s}){Pn();var c=void 0,l=Xn(()=>{var o=n??t.appendChild(R());an(o,{pending:()=>{}},t=>{wt({});var n=k;if(a&&(n.c=a),i&&(r.$$events=i),T&&Kr(t,null),c=e(t,r)||{},T&&(G.nodes.end=E,E===null||E.nodeType!==8||E.data!==`]`))throw lt(),st;Tt()},s);var l=new Set,u=e=>{for(var n=0;n<e.length;n++){var r=e[n];if(!l.has(r)){l.add(r);var i=Pr(r);for(let e of[t,document]){var a=Yr.get(e);a===void 0&&(a=new Map,Yr.set(e,a));var o=a.get(r);o===void 0?(e.addEventListener(r,Hr,{passive:i}),a.set(r,1)):a.set(r,o+1)}}}};return u(ve(Ir)),Lr.add(u),()=>{for(var e of l)for(let n of[t,document]){var r=Yr.get(n),i=r.get(e);--i==0?(n.removeEventListener(e,Hr),r.delete(e),r.size===0&&Yr.delete(n)):r.set(e,i)}Lr.delete(u),o!==n&&o.parentNode?.removeChild(o)}});return Zr.set(c,l),c}var Zr=new WeakMap;function Qr(e,t){let n=Zr.get(e);return n?(Zr.delete(e),n(t)):Promise.resolve()}var $r=class{anchor;#e=new Map;#t=new Map;#n=new Map;#r=new Set;#i=!0;constructor(e,t=!0){this.anchor=e,this.#i=t}#a=e=>{if(this.#e.has(e)){var t=this.#e.get(e),n=this.#t.get(t);if(n)lr(n),this.#r.delete(t);else{var r=this.#n.get(t);r&&(this.#t.set(t,r.effect),this.#n.delete(t),r.fragment.lastChild.remove(),this.anchor.before(r.fragment),n=r.effect)}for(let[t,n]of this.#e){if(this.#e.delete(t),t===e)break;let r=this.#n.get(n);r&&(V(r.effect),this.#n.delete(n))}for(let[e,r]of this.#t){if(e===t||this.#r.has(e))continue;let i=()=>{if(Array.from(this.#e.values()).includes(e)){var t=document.createDocumentFragment();dr(r,t),t.append(R()),this.#n.set(e,{effect:r,fragment:t})}else V(r);this.#r.delete(e),this.#t.delete(e)};this.#i||!n?(this.#r.add(e),sr(r,i,!1)):i()}}};#o=e=>{this.#e.delete(e);let t=Array.from(this.#e.values());for(let[e,n]of this.#n)t.includes(e)||(V(n.effect),this.#n.delete(e))};ensure(e,t){var n=M,r=zn();if(t&&!this.#t.has(e)&&!this.#n.has(e))if(r){var i=document.createDocumentFragment(),a=R();i.append(a),this.#n.set(e,{effect:B(()=>t(a)),fragment:i})}else this.#t.set(e,B(()=>t(this.anchor)));if(this.#e.set(n,e),r){for(let[t,r]of this.#t)t===e?n.unskip_effect(r):n.skip_effect(r);for(let[t,r]of this.#n)t===e?n.unskip_effect(r.effect):n.skip_effect(r.effect);n.oncommit(this.#a),n.ondiscard(this.#o)}else T&&(this.anchor=E),this.#a(n)}};function ei(e,t,n=!1){var r;T&&(r=E,ft());var i=new $r(e),a=n?Fe:0;function o(e,t){if(T){var n=ht(r);if(e!==parseInt(n.substring(1))){var a=mt();D(a),i.anchor=a,dt(!1),i.ensure(e,t),dt(!0);return}}i.ensure(e,t)}tr(()=>{var e=!1;t((t,n=0)=>{e=!0,o(n,t)}),e||o(-1,null)},a)}function ti(e,t){return t}function ni(e,t,n){for(var r=[],i=t.length,a,o=t.length,s=0;s<i;s++){let n=t[s];sr(n,()=>{if(a){if(a.pending.delete(n),a.done.add(n),a.pending.size===0){var t=e.outrogroups;ri(e,ve(a.done)),t.delete(a),t.size===0&&(e.outrogroups=null)}}else --o},!1)}if(o===0){var c=r.length===0&&n!==null;if(c){var l=n,u=l.parentNode;Rn(u),u.append(l),e.items.clear()}ri(e,t,!c)}else a={pending:new Set(t),done:new Set},(e.outrogroups??=new Set).add(a)}function ri(e,t,n=!0){var r;if(e.pending.size>0){r=new Set;for(let t of e.pending.values())for(let n of t)r.add(e.items.get(n).e)}for(var i=0;i<t.length;i++){var a=t[i];r?.has(a)?(a.f|=Re,dr(a,document.createDocumentFragment())):V(t[i],n)}}var ii;function ai(e,t,n,r,i,a=null){var o=e,s=new Map;if(t&4){var c=e;o=T?D(Fn(c)):c.appendChild(R())}T&&ft();var l=null,u=pn(()=>{var e=n();return he(e)?e:e==null?[]:ve(e)}),d,f=new Map,p=!0;function m(e){g.effect.f&16384||(g.pending.delete(e),g.fallback=l,si(g,d,o,t,r),l!==null&&(d.length===0?l.f&33554432?(l.f^=Re,li(l,null,o)):lr(l):sr(l,()=>{l=null})))}function h(e){g.pending.delete(e)}var g={effect:tr(()=>{d=X(u);var e=d.length;let c=!1;T&&ht(o)===`[!`!=(e===0)&&(o=mt(),D(o),dt(!1),c=!0);for(var g=new Set,_=M,v=zn(),y=0;y<e;y+=1){T&&E.nodeType===8&&E.data===`]`&&(o=E,c=!0,dt(!1));var b=d[y],ee=r(b,y),x=p?null:s.get(ee);x?(x.v&&Tn(x.v,b),x.i&&Tn(x.i,y),v&&_.unskip_effect(x.e)):(x=ci(s,p?o:ii??=R(),b,ee,y,i,t,n),p||(x.e.f|=Re),s.set(ee,x)),g.add(ee)}if(e===0&&a&&!l&&(p?l=B(()=>a(o)):(l=B(()=>a(ii??=R())),l.f|=Re)),e>g.size&&Xe(``,``,``),T&&e>0&&D(mt()),!p)if(f.set(_,g),v){for(let[e,t]of s)g.has(e)||_.skip_effect(t.e);_.oncommit(m),_.ondiscard(h)}else m(_);c&&dt(!0),X(u)}),flags:t,items:s,pending:f,outrogroups:null,fallback:l};p=!1,T&&(o=E)}function oi(e){for(;e!==null&&!(e.f&32);)e=e.next;return e}function si(e,t,n,r,i){var a=(r&8)!=0,o=t.length,s=e.items,c=oi(e.effect.first),l,u=null,d,f=[],p=[],m,h,g,_;if(a)for(_=0;_<o;_+=1)m=t[_],h=i(m,_),g=s.get(h).e,g.f&33554432||(g.nodes?.a?.measure(),(d??=new Set).add(g));for(_=0;_<o;_+=1){if(m=t[_],h=i(m,_),g=s.get(h).e,e.outrogroups!==null)for(let t of e.outrogroups)t.pending.delete(g),t.done.delete(g);if(g.f&8192&&(lr(g),a&&(g.nodes?.a?.unfix(),(d??=new Set).delete(g))),g.f&33554432)if(g.f^=Re,g===c)li(g,null,n);else{var v=u?u.next:c;g===e.effect.last&&(e.effect.last=g.prev),g.prev&&(g.prev.next=g.next),g.next&&(g.next.prev=g.prev),ui(e,u,g),ui(e,g,v),li(g,v,n),u=g,f=[],p=[],c=oi(u.next);continue}if(g!==c){if(l!==void 0&&l.has(g)){if(f.length<p.length){var y=p[0],b;u=y.prev;var ee=f[0],x=f[f.length-1];for(b=0;b<f.length;b+=1)li(f[b],y,n);for(b=0;b<p.length;b+=1)l.delete(p[b]);ui(e,ee.prev,x.next),ui(e,u,ee),ui(e,x,y),c=y,u=x,--_,f=[],p=[]}else l.delete(g),li(g,c,n),ui(e,g.prev,g.next),ui(e,g,u===null?e.effect.first:u.next),ui(e,u,g),u=g;continue}for(f=[],p=[];c!==null&&c!==g;)(l??=new Set).add(c),p.push(c),c=oi(c.next);if(c===null)continue}g.f&33554432||f.push(g),u=g,c=oi(g.next)}if(e.outrogroups!==null){for(let t of e.outrogroups)t.pending.size===0&&(ri(e,ve(t.done)),e.outrogroups?.delete(t));e.outrogroups.size===0&&(e.outrogroups=null)}if(c!==null||l!==void 0){var te=[];if(l!==void 0)for(g of l)g.f&8192||te.push(g);for(;c!==null;)!(c.f&8192)&&c!==e.fallback&&te.push(c),c=oi(c.next);var ne=te.length;if(ne>0){var re=r&4&&o===0?n:null;if(a){for(_=0;_<ne;_+=1)te[_].nodes?.a?.measure();for(_=0;_<ne;_+=1)te[_].nodes?.a?.fix()}ni(e,te,re)}}a&&jt(()=>{if(d!==void 0)for(g of d)g.nodes?.a?.apply()})}function ci(e,t,n,r,i,a,o,s){var c=o&1?o&16?Cn(n):wn(n,!1,!1):null,l=o&2?Cn(i):null;return{v:c,i:l,e:B(()=>(a(t,c??n,l??i,s),()=>{e.delete(r)}))}}function li(e,t,n){if(e.nodes)for(var r=e.nodes.start,i=e.nodes.end,a=t&&!(t.f&33554432)?t.nodes.start:n;r!==null;){var o=In(r);if(a.before(r),r===i)return;r=o}}function ui(e,t,n){t===null?e.effect.first=n:t.next=n,n===null?e.effect.last=t:n.prev=t}function di(e,t){Zn(()=>{var n=e.getRootNode(),r=n.host?n:n.head??n.ownerDocument.head;if(!r.querySelector(`#`+t.hash)){let e=Bn(`style`);e.id=t.hash,e.textContent=t.code,r.appendChild(e)}})}var fi=[...` 	
+\r\f\xA0\v﻿`];function pi(e,t,n){var r=e==null?``:``+e;if(t&&(r=r?r+` `+t:t),n){for(var i of Object.keys(n))if(n[i])r=r?r+` `+i:i;else if(r.length)for(var a=i.length,o=0;(o=r.indexOf(i,o))>=0;){var s=o+a;(o===0||fi.includes(r[o-1]))&&(s===r.length||fi.includes(r[s]))?r=(o===0?``:r.substring(0,o))+r.substring(s+1):o=s}}return r===``?null:r}function mi(e,t,n,r,i,a){var o=e.__className;if(T||o!==n||o===void 0){var s=pi(n,r,a);(!T||s!==e.getAttribute(`class`))&&(s==null?e.removeAttribute(`class`):t?e.className=s:e.setAttribute(`class`,s)),e.__className=n}else if(a&&i!==a)for(var c in a){var l=!!a[c];(i==null||l!==!!i[c])&&e.classList.toggle(c,l)}return a}var hi=Symbol(`is custom element`),gi=Symbol(`is html`),_i=qe?`link`:`LINK`;function vi(e,t,n,r){var i=yi(e);T&&(i[t]=e.getAttribute(t),t===`src`||t===`srcset`||t===`href`&&e.nodeName===_i)||i[t]!==(i[t]=n)&&(t===`loading`&&(e[Ge]=n),n==null?e.removeAttribute(t):typeof n!=`string`&&xi(e).includes(t)?e[t]=n:e.setAttribute(t,n))}function yi(e){return e.__attributes??={[hi]:e.nodeName.includes(`-`),[gi]:e.namespaceURI===ct}}var bi=new Map;function xi(e){var t=e.getAttribute(`is`)||e.nodeName,n=bi.get(t);if(n)return n;bi.set(t,n=[]);for(var r,i=e,a=Element.prototype;a!==i;){for(var o in r=xe(i),r)r[o].set&&n.push(o);i=we(i)}return n}var Si={get(e,t){let n=e.props.length;for(;n--;){let r=e.props[n];if(Ee(r)&&(r=r()),typeof r==`object`&&r&&t in r)return r[t]}},set(e,t,n){let r=e.props.length;for(;r--;){let i=e.props[r];Ee(i)&&(i=i());let a=be(i,t);if(a&&a.set)return a.set(n),!0}return!1},getOwnPropertyDescriptor(e,t){let n=e.props.length;for(;n--;){let r=e.props[n];if(Ee(r)&&(r=r()),typeof r==`object`&&r&&t in r){let e=be(r,t);return e&&!e.configurable&&(e.configurable=!0),e}}},has(e,t){if(t===Ue||t===We)return!1;for(let n of e.props)if(Ee(n)&&(n=n()),n!=null&&t in n)return!0;return!1},ownKeys(e){let t=[];for(let n of e.props)if(Ee(n)&&(n=n()),n){for(let e in n)t.includes(e)||t.push(e);for(let e of Object.getOwnPropertySymbols(n))t.includes(e)||t.push(e)}return t}};function Ci(...e){return new Proxy({props:e},Si)}function wi(e,t,n,r){var i=!bt||(n&2)!=0,a=(n&8)!=0,o=(n&16)!=0,s=r,c=!0,l=()=>(c&&(c=!1,s=o?Mr(r):r),s);let u;if(a){var d=Ue in e||We in e;u=be(e,t)?.set??(d&&t in e?n=>e[t]=n:void 0)}var f,p=!1;a?[f,p]=Bt(()=>e[t]):f=e[t],f===void 0&&r!==void 0&&(f=l(),u&&(i&&tt(t),u(f)));var m=i?()=>{var n=e[t];return n===void 0?l():(c=!0,n)}:()=>{var n=e[t];return n!==void 0&&(s=void 0),n===void 0?s:n};if(i&&!(n&4))return m;if(u){var h=e.$$legacy;return(function(e,t){return arguments.length>0?((!i||!t||h||p)&&u(t?m():e),e):m()})}var g=!1,_=(n&1?dn:pn)(()=>(g=!1,m()));a&&X(_);var v=G;return(function(e,t){if(arguments.length>0){let n=t?X(_):i&&a?kn(e):e;return L(_,n),g=!0,s!==void 0&&(s=n),e}return mr&&g||v.f&16384?_.v:X(_)})}typeof window<`u`&&((window.__svelte??={}).v??=new Set).add(`5`);function Ti(t){return e.Platform.isMacOS?t.metaKey:t.ctrlKey}function Ei(e,t){return`${e}:${t.format(a[e])}`}function Di(e,t,n){let r=new Map,i=e[1].days[0];for(let i of e){for(let e of i.days)r.set(Ei(`day`,e),t(e,`day`));if(n.includes(`week`)){let e=i.days[0];r.set(Ei(`week`,e),t(e,`week`))}}return n.includes(`month`)&&r.set(Ei(`month`,i),t(i,`month`)),n.includes(`year`)&&r.set(Ei(`year`,i),t(i,`year`)),r}var Oi=Q(`<td><div role="button" tabindex="0"> </div></td>`),ki={hash:`svelte-1l5s8qw`,code:`.day.svelte-1l5s8qw {background-color:var(--color-background-day);border-radius:4px;color:var(--color-text-day);cursor:pointer;font-size:0.8em;height:100%;padding:4px;position:relative;text-align:center;transition:background-color 0.1s ease-in,
       color 0.1s ease-in;vertical-align:baseline;}.day.svelte-1l5s8qw:hover {background-color:var(--interactive-hover);}.day.active.svelte-1l5s8qw:hover {background-color:var(--interactive-accent-hover);}.adjacent-month.svelte-1l5s8qw {opacity:0.25;}.has-note.svelte-1l5s8qw::after {background-color:var(--text-muted);border-radius:50%;content:"";display:block;height:3px;margin:1px auto 0;width:3px;}.has-note.active.svelte-1l5s8qw::after {background-color:var(--text-on-accent);}.today.svelte-1l5s8qw {color:var(--interactive-accent);font-weight:600;}.day.svelte-1l5s8qw:active,
   .active.svelte-1l5s8qw,
-  .active.today.svelte-1l5s8qw {color:var(--text-on-accent);background-color:var(--interactive-accent);}`};function Ni(e,t){wt(t,!0),pi(e,Mi);let n=Ei(t,`activeFilePath`,3,null),r=St(u),i=P(()=>t.fileMap.get(ki(`day`,t.date))??null);function a(e){t.onClick?.(`day`,t.date,X(i),Di(e))}function o(e){e.target&&t.onHover?.(`day`,t.date,X(i),e.target,Di(e))}function s(e){t.onContextMenu?.(`day`,t.date,X(i),e)}var c=ji(),l=R(c);let d;var f=R(l,!0);O(l),O(c),nr((e,t)=>{d=gi(l,1,`day svelte-1l5s8qw`,null,d,e),Yr(f,t)},[()=>({active:X(i)!==null&&X(i).path===n(),"adjacent-month":!t.date.isSame(r.current,`month`),"has-note":X(i)!==null,today:t.date.isSame(t.today,`day`)}),()=>t.date.format(`D`)]),Z(`click`,l,a),Z(`keydown`,l,e=>{(e.key===`Enter`||e.key===` `)&&(e.preventDefault(),t.onClick?.(`day`,t.date,X(i),!1))}),Z(`contextmenu`,l,s),Vr(`pointerenter`,l,o),$(e,c),Tt()}Hr([`click`,`keydown`,`contextmenu`]);var Pi=class{#e=F(window.moment());get current(){return X(this.#e)}set current(e){I(this.#e,e)}},Fi=Q(`<button type="button"><svg focusable="false" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 320 512" class="svelte-he81r3"><path fill="currentColor" d="M34.52 239.03L228.87 44.69c9.37-9.37 24.57-9.37 33.94 0l22.67 22.67c9.36 9.36 9.37 24.52.04 33.9L131.49 256l154.02 154.75c9.34 9.38 9.32 24.54-.04 33.9l-22.67 22.67c-9.37 9.37-24.57 9.37-33.94 0L34.52 272.97c-9.37-9.37-9.37-24.57 0-33.94z"></path></svg></button>`),Ii={hash:`svelte-he81r3`,code:`.arrow.svelte-he81r3 {align-items:center;appearance:none;background:none;border:none;cursor:pointer;display:flex;justify-content:center;padding:0;width:24px;}.arrow.is-mobile.svelte-he81r3 {width:32px;}.right.svelte-he81r3 {transform:rotate(180deg);}.arrow.svelte-he81r3 svg:where(.svelte-he81r3) {color:var(--color-arrow);height:16px;width:16px;}`};function Li(t,n){wt(n,!0),pi(t,Ii);var r=Fi();let i;nr(()=>{i=gi(r,1,`arrow svelte-he81r3`,null,i,{"is-mobile":e.Platform.isMobile,right:n.direction===`right`}),bi(r,`aria-label`,n.tooltip)}),Z(`click`,r,function(...e){n.onClick?.apply(this,e)}),$(t,r),Tt()}Hr([`click`]);var Ri=Q(`<div><span class="title svelte-16hzpd8"><span> </span> <span> </span></span></div>`),zi={hash:`svelte-16hzpd8`,code:`.title.svelte-16hzpd8 {color:var(--color-text-title);display:flex;font-size:1.4em;gap:0.3em;margin:0;}.month.svelte-16hzpd8 {font-weight:500;}.year.svelte-16hzpd8 {color:var(--interactive-accent);}.clickable.svelte-16hzpd8 {cursor:pointer;}`};function Bi(e,t){wt(t,!0),pi(e,zi);let n=St(u),r=P(()=>ki(`month`,n.current)),i=P(()=>ki(`year`,n.current)),a=P(()=>t.fileMap.has(X(r))),o=P(()=>t.fileMap.has(X(i))),s=P(()=>t.fileMap.get(X(r))??null),c=P(()=>t.fileMap.get(X(i))??null);function l(e,r,i){return{click:a=>{r()?t.onClick?.(e,n.current,i(),Di(a)):e===`month`&&t.resetDisplayedMonth()},hover:a=>{!r()||!a.target||t.onHover?.(e,n.current,i(),a.target,Di(a))},context:a=>{let o=i();r()&&o&&t.onContextMenu?.(e,n.current,o,a)}}}let d=l(`month`,()=>X(a),()=>X(s)),f=l(`year`,()=>X(o),()=>X(c));var p=Ri(),m=R(p),h=R(m);let g;var _=R(h,!0);O(h);var v=z(h,2);let y;var b=R(v,!0);O(v),O(m),O(p),nr((e,t)=>{g=gi(h,1,`month svelte-16hzpd8`,null,g,{clickable:X(a)}),bi(h,`role`,X(a)?`button`:void 0),bi(h,`tabindex`,X(a)?0:void 0),Yr(_,e),y=gi(v,1,`year svelte-16hzpd8`,null,y,{clickable:X(o)}),bi(v,`role`,X(o)?`button`:void 0),bi(v,`tabindex`,X(o)?0:void 0),Yr(b,t)},[()=>n.current.format(`MMM`),()=>n.current.format(`YYYY`)]),Z(`click`,h,function(...e){d.click?.apply(this,e)}),Z(`keydown`,h,e=>{(e.key===`Enter`||e.key===` `)&&(X(a)?t.onClick?.(`month`,n.current,X(s),!1):t.resetDisplayedMonth())}),Z(`contextmenu`,h,function(...e){d.context?.apply(this,e)}),Vr(`pointerenter`,h,function(...e){d.hover?.apply(this,e)}),Z(`click`,v,function(...e){f.click?.apply(this,e)}),Z(`keydown`,v,e=>{(e.key===`Enter`||e.key===` `)&&X(o)&&t.onClick?.(`year`,n.current,X(c),!1)}),Z(`contextmenu`,v,function(...e){f.context?.apply(this,e)}),Vr(`pointerenter`,v,function(...e){f.hover?.apply(this,e)}),$(e,p),Tt()}Hr([`click`,`keydown`,`contextmenu`]);var Vi=Q(`<div class="nav svelte-17quior"><!> <div class="right-nav svelte-17quior"><!> <button type="button">&#x25CF;</button> <!></div></div>`),Hi={hash:`svelte-17quior`,code:`.nav.svelte-17quior {align-items:baseline;display:flex;margin:0.6em 0 1em;padding:0 8px;width:100%;}.right-nav.svelte-17quior {align-items:center;display:flex;justify-content:center;margin-left:auto;}.reset-button.svelte-17quior {align-items:center;appearance:none;background:none;border:none;color:var(--color-arrow);cursor:default;display:flex;font-size:8px;opacity:0.4;padding:0.5em;}.reset-button.active.svelte-17quior {cursor:pointer;opacity:1;}`};function Ui(e,t){wt(t,!0),pi(e,Hi);let n=St(u);function r(){n.current=n.current.clone().add(1,`month`)}function i(){n.current=n.current.clone().subtract(1,`month`)}function a(){n.current=t.today.clone()}let o=P(()=>n.current.isSame(t.today,`month`));var s=Vi(),c=R(s);Bi(c,Ti({get fileMap(){return t.fileMap},resetDisplayedMonth:a},()=>t.eventHandlers));var l=z(c,2),d=R(l);Li(d,{direction:`left`,onClick:i,tooltip:`Previous Month`});var f=z(d,2);let p;Li(z(f,2),{direction:`right`,onClick:r,tooltip:`Next Month`}),O(l),O(s),nr(()=>{bi(f,`aria-label`,X(o)?`Current month`:`Reset to current month`),p=gi(f,1,`reset-button svelte-17quior`,null,p,{active:!X(o)})}),Z(`click`,f,a),$(e,s),Tt()}Hr([`click`]);function Wi(){return window.moment.weekdaysShort(!0)}function Gi(e){return e.isoWeekday()===6||e.isoWeekday()===7}function Ki(e){return e[0].clone()}function qi(e){let t=[],n,r=e.clone().date(1),i=r.weekday(),a=r.clone().subtract(i,`days`);for(let e=0;e<42;e++)e%7==0&&(n={days:[],weekNum:a.week()},t.push(n)),n.days.push(a),a=a.clone().add(1,`days`);return t}var Ji=Q(`<td class="svelte-uujyui"><div role="button" tabindex="0"> </div></td>`),Yi={hash:`svelte-uujyui`,code:`td.svelte-uujyui {border-right:1px solid var(--background-modifier-border);}.week-num.svelte-uujyui {background-color:var(--color-background-weeknum);border-radius:4px;color:var(--color-text-weeknum);cursor:pointer;font-size:0.65em;height:100%;padding:4px;text-align:center;transition:background-color 0.1s ease-in,
-      color 0.1s ease-in;vertical-align:baseline;}.week-num.svelte-uujyui:hover {background-color:var(--interactive-hover);}.week-num.active.svelte-uujyui:hover {background-color:var(--interactive-accent-hover);}.active.svelte-uujyui {color:var(--text-on-accent);background-color:var(--interactive-accent);}`};function Xi(e,t){wt(t,!0),pi(e,Yi);let n=Ei(t,`activeFilePath`,3,null),r=P(()=>Ki(t.days)),i=P(()=>t.fileMap.get(ki(`week`,X(r)))??null);function a(e){e.target&&t.onHover?.(`week`,X(r),X(i),e.target,Di(e))}var o=Ji(),s=R(o);let c;var l=R(s,!0);O(s),O(o),nr(()=>{c=gi(s,1,`week-num svelte-uujyui`,null,c,{active:X(i)!==null&&X(i).path===n()}),Yr(l,t.weekNum)}),Z(`click`,s,function(...e){(t.onClick&&(e=>t.onClick(`week`,X(r),X(i),Di(e))))?.apply(this,e)}),Z(`keydown`,s,function(...e){(t.onClick&&(e=>(e.key===`Enter`||e.key===` `)&&t.onClick(`week`,X(r),X(i),!1)))?.apply(this,e)}),Z(`contextmenu`,s,function(...e){(t.onContextMenu&&(e=>t.onContextMenu(`week`,X(r),X(i),e)))?.apply(this,e)}),Vr(`pointerenter`,s,a),$(e,o),Tt()}Hr([`click`,`keydown`,`contextmenu`]);var Zi=Q(`<col/>`),Qi=Q(`<col/>`),$i=Q(`<th class="svelte-5j20fs">W</th>`),ea=Q(`<th class="svelte-5j20fs"> </th>`),ta=Q(`<tr><!><!></tr>`),na=Q(`<div id="calendar-container" class="container svelte-5j20fs"><!> <table class="calendar svelte-5j20fs"><colgroup><!><!></colgroup><thead><tr><!><!></tr></thead><tbody></tbody></table></div>`),ra={hash:`svelte-5j20fs`,code:`.container.svelte-5j20fs {--color-background-heading: transparent;--color-background-day: transparent;--color-background-weeknum: transparent;--color-background-weekend: transparent;--color-arrow: var(--text-muted);--color-button: var(--text-muted);--color-text-title: var(--text-normal);--color-text-heading: var(--text-muted);--color-text-day: var(--text-normal);--color-text-today: var(--interactive-accent);--color-text-weeknum: var(--text-muted);}.container.svelte-5j20fs {padding:0 8px;}.weekend.svelte-5j20fs {background-color:var(--color-background-weekend);}.calendar.svelte-5j20fs {border-collapse:collapse;width:100%;}th.svelte-5j20fs {background-color:var(--color-background-heading);color:var(--color-text-heading);font-size:0.6em;letter-spacing:1px;padding:4px;text-align:center;text-transform:uppercase;}`};function ia(e,t){wt(t,!0),pi(e,ra);let n=F(null),r=F(window.moment()),i=new Pi;Ct(u,i);let a=F(qi(window.moment())),o=F(!1),s=F(new Map);Xn(()=>{I(a,qi(i.current))}),Xn(()=>{let e=X(a);return t.fileStore.store.subscribe(()=>{I(o,t.fileStore.isGranularityEnabled(`week`),!0),I(s,Ai(e,(e,n)=>t.fileStore.getFile(e,n),t.fileStore.getEnabledGranularities()))})});let c=P(()=>({onHover:t.onHover,onClick:t.onClick,onContextMenu:t.onContextMenu})),l=Wi();function d(){let e=window.moment();e.isSame(X(r),`day`)||I(r,e)}function f(e){I(n,e,!0)}var p={tick:d,setActiveFilePath:f},m=na(),h=R(m);Ui(h,{get fileMap(){return X(s)},get today(){return X(r)},get eventHandlers(){return X(c)}});var g=z(h,2),_=R(g),v=R(_),y=e=>{$(e,Zi())};ni(v,e=>{X(o)&&e(y)}),si(z(v),17,()=>X(a)[1].days,ri,(e,t)=>{var n=Qi();let r;nr(e=>r=gi(n,1,`svelte-5j20fs`,null,r,e),[()=>({weekend:Gi(X(t))})]),$(e,n)}),O(_);var b=z(_),ee=R(b),x=R(ee),te=e=>{$(e,$i())};ni(x,e=>{X(o)&&e(te)}),si(z(x),17,()=>l,ri,(e,t)=>{var n=ea(),r=R(n,!0);O(n),nr(()=>Yr(r,X(t))),$(e,n)}),O(ee),O(b);var ne=z(b);return si(ne,21,()=>X(a),e=>ki(`week`,e.days[0]),(e,t)=>{var i=ta(),a=R(i),l=e=>{Xi(e,Ti({get fileMap(){return X(s)},get activeFilePath(){return X(n)}},()=>X(t),()=>X(c)))};ni(a,e=>{X(o)&&e(l)}),si(z(a),17,()=>X(t).days,e=>e.format(),(e,t)=>{Ni(e,Ti({get date(){return X(t)},get fileMap(){return X(s)},get today(){return X(r)},get activeFilePath(){return X(n)}},()=>X(c)))}),O(i),$(e,i)}),O(ne),O(g),O(m),$(e,m),Tt(p)}var aa=class extends e.ItemView{calendar;plugin;constructor(e,t){super(e),this.plugin=t,this.registerEvent(this.app.workspace.on(`file-open`,this.onFileOpen.bind(this)))}getViewType(){return l}getDisplayText(){return`Calendar`}getIcon(){return`calendar-day`}async onClose(){this.calendar&&ei(this.calendar)}async onOpen(){let e=new Oi(this,this.plugin),t=Xr(ia,{target:this.contentEl,props:{fileStore:e,onHover:this.onHover.bind(this),onClick:this.onClick.bind(this),onContextMenu:this.onContextMenu.bind(this)}});if(!(`tick`in t&&`setActiveFilePath`in t))throw Error(`Calendar component missing expected exports`);this.calendar=t}onHover(e,t,n,r,i){if(!i)return;let a=t.format(e===`day`?`YYYY-MM-DD`:t.localeData().longDateFormat(`L`));this.app.workspace.trigger(`link-hover`,this,r,a,n?.path??``)}onClick(e,t,n,r){this.plugin.openPeriodicNote(e,t,{inNewSplit:r})}onContextMenu(t,n,r,i){if(!r)return;let a=new e.Menu;a.addItem(e=>e.setTitle(`Delete`).setIcon(`trash`).onClick(()=>{this.app.vault.trash(r,!0)})),this.app.workspace.trigger(`file-menu`,a,r,`calendar-context-menu`,null),a.showAtPosition({x:i.pageX,y:i.pageY})}onFileOpen(e){if(this.app.workspace.layoutReady&&this.calendar){let e=this.app.workspace.getActiveFile()?.path??null;this.calendar.setActiveFilePath(e),this.calendar.tick()}}},oa={day:{periodicity:`daily`,relativeUnit:`today`,labelOpenPresent:`Open today's daily note`},week:{periodicity:`weekly`,relativeUnit:`this week`,labelOpenPresent:`Open this week's note`},month:{periodicity:`monthly`,relativeUnit:`this month`,labelOpenPresent:`Open this month's note`},year:{periodicity:`yearly`,relativeUnit:`this year`,labelOpenPresent:`Open this year's note`}};async function sa(t,n,r){let i=t.workspace.getActiveFile();if(!i)return;let a=n.findInCache(i.path);if(!a)return;let o=n.findAdjacent(i.path,r);if(o){let n=t.vault.getAbstractFileByPath(o.filePath);n&&n instanceof e.TFile&&await t.workspace.getLeaf().openFile(n,{active:!0})}else{let t=r===`forwards`?`after`:`before`;new e.Notice(`There's no ${oa[a.granularity].periodicity} note ${t} this`)}}async function ca(e,t,n){let r=e.workspace.getActiveFile();if(!r)return;let i=t.findInCache(r.path);if(!i)return;let a=n===`forwards`?1:-1,o=i.date.clone().add(a,i.granularity);t.openPeriodicNote(i.granularity,o)}function la(e,t,n){let r=oa[n];return[{id:`open-${r.periodicity}-note`,name:r.labelOpenPresent,checkCallback:e=>{if(!t.settings.granularities[n].enabled)return!1;if(e)return!0;t.openPeriodicNote(n,window.moment())}},{id:`next-${r.periodicity}-note`,name:`Jump forwards to closest ${r.periodicity} note`,checkCallback:r=>{if(!t.settings.granularities[n].enabled)return!1;let i=e.workspace.getActiveFile();if(r)return i?t.isPeriodic(i.path,n):!1;sa(e,t,`forwards`)}},{id:`prev-${r.periodicity}-note`,name:`Jump backwards to closest ${r.periodicity} note`,checkCallback:r=>{if(!t.settings.granularities[n].enabled)return!1;let i=e.workspace.getActiveFile();if(r)return i?t.isPeriodic(i.path,n):!1;sa(e,t,`backwards`)}},{id:`open-next-${r.periodicity}-note`,name:`Open next ${r.periodicity} note`,checkCallback:r=>{if(!t.settings.granularities[n].enabled)return!1;let i=e.workspace.getActiveFile();if(r)return i?t.isPeriodic(i.path,n):!1;ca(e,t,`forwards`)}},{id:`open-prev-${r.periodicity}-note`,name:`Open previous ${r.periodicity} note`,checkCallback:r=>{if(!t.settings.granularities[n].enabled)return!1;let i=e.workspace.getActiveFile();if(r)return i?t.isPeriodic(i.path,n):!1;ca(e,t,`backwards`)}}]}function ua(t,n){let i=new e.Menu,a=r.filter(e=>t.settings.granularities[e].enabled);for(let e of a){let n=oa[e];i.addItem(r=>r.setTitle(n.labelOpenPresent).setIcon(`calendar-${e}`).onClick(()=>t.openPeriodicNote(e,window.moment())))}i.showAtPosition(n)}var da=`
+  .active.today.svelte-1l5s8qw {color:var(--text-on-accent);background-color:var(--interactive-accent);}`};function Ai(e,t){wt(t,!0),di(e,ki);let n=wi(t,`activeFilePath`,3,null),r=St(u),i=F(()=>t.fileMap.get(Ei(`day`,t.date))??null);function a(e){t.onClick?.(`day`,t.date,X(i),Ti(e))}function o(e){e.target&&t.onHover?.(`day`,t.date,X(i),e.target,Ti(e))}function s(e){t.onContextMenu?.(`day`,t.date,X(i),e)}var c=Oi(),l=z(c);let d;var f=z(l,!0);O(l),O(c),er((e,t)=>{d=mi(l,1,`day svelte-1l5s8qw`,null,d,e),qr(f,t)},[()=>({active:X(i)!==null&&X(i).path===n(),"adjacent-month":!t.date.isSame(r.current,`month`),"has-note":X(i)!==null,today:t.date.isSame(t.today,`day`)}),()=>t.date.format(`D`)]),Z(`click`,l,a),Z(`keydown`,l,e=>{(e.key===`Enter`||e.key===` `)&&(e.preventDefault(),t.onClick?.(`day`,t.date,X(i),!1))}),Z(`contextmenu`,l,s),zr(`pointerenter`,l,o),$(e,c),Tt()}Br([`click`,`keydown`,`contextmenu`]);var ji=class{#e=I(window.moment());get current(){return X(this.#e)}set current(e){L(this.#e,e)}},Mi=Q(`<button type="button"><svg focusable="false" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 320 512" class="svelte-he81r3"><path fill="currentColor" d="M34.52 239.03L228.87 44.69c9.37-9.37 24.57-9.37 33.94 0l22.67 22.67c9.36 9.36 9.37 24.52.04 33.9L131.49 256l154.02 154.75c9.34 9.38 9.32 24.54-.04 33.9l-22.67 22.67c-9.37 9.37-24.57 9.37-33.94 0L34.52 272.97c-9.37-9.37-9.37-24.57 0-33.94z"></path></svg></button>`),Ni={hash:`svelte-he81r3`,code:`.arrow.svelte-he81r3 {align-items:center;appearance:none;background:none;border:none;cursor:pointer;display:flex;justify-content:center;padding:0;width:24px;}.arrow.is-mobile.svelte-he81r3 {width:32px;}.right.svelte-he81r3 {transform:rotate(180deg);}.arrow.svelte-he81r3 svg:where(.svelte-he81r3) {color:var(--color-arrow);height:16px;width:16px;}`};function Pi(t,n){wt(n,!0),di(t,Ni);var r=Mi();let i;er(()=>{i=mi(r,1,`arrow svelte-he81r3`,null,i,{"is-mobile":e.Platform.isMobile,right:n.direction===`right`}),vi(r,`aria-label`,n.tooltip)}),Z(`click`,r,function(...e){n.onClick?.apply(this,e)}),$(t,r),Tt()}Br([`click`]);var Fi=Q(`<div><span class="title svelte-16hzpd8"><span> </span> <span> </span></span></div>`),Ii={hash:`svelte-16hzpd8`,code:`.title.svelte-16hzpd8 {color:var(--color-text-title);display:flex;font-size:1.4em;gap:0.3em;margin:0;}.month.svelte-16hzpd8 {font-weight:500;}.year.svelte-16hzpd8 {color:var(--interactive-accent);}.clickable.svelte-16hzpd8 {cursor:pointer;}`};function Li(e,t){wt(t,!0),di(e,Ii);let n=St(u),r=F(()=>Ei(`month`,n.current)),i=F(()=>Ei(`year`,n.current)),a=F(()=>t.fileMap.has(X(r))),o=F(()=>t.fileMap.has(X(i))),s=F(()=>t.fileMap.get(X(r))??null),c=F(()=>t.fileMap.get(X(i))??null);function l(e,r,i){return{click:a=>{r()?t.onClick?.(e,n.current,i(),Ti(a)):e===`month`&&t.resetDisplayedMonth()},hover:a=>{!r()||!a.target||t.onHover?.(e,n.current,i(),a.target,Ti(a))},context:a=>{let o=i();r()&&o&&t.onContextMenu?.(e,n.current,o,a)}}}let d=l(`month`,()=>X(a),()=>X(s)),f=l(`year`,()=>X(o),()=>X(c));var p=Fi(),m=z(p),h=z(m);let g;var _=z(h,!0);O(h);var v=Ln(h,2);let y;var b=z(v,!0);O(v),O(m),O(p),er((e,t)=>{g=mi(h,1,`month svelte-16hzpd8`,null,g,{clickable:X(a)}),vi(h,`role`,X(a)?`button`:void 0),vi(h,`tabindex`,X(a)?0:void 0),qr(_,e),y=mi(v,1,`year svelte-16hzpd8`,null,y,{clickable:X(o)}),vi(v,`role`,X(o)?`button`:void 0),vi(v,`tabindex`,X(o)?0:void 0),qr(b,t)},[()=>n.current.format(`MMM`),()=>n.current.format(`YYYY`)]),Z(`click`,h,function(...e){d.click?.apply(this,e)}),Z(`keydown`,h,e=>{(e.key===`Enter`||e.key===` `)&&(X(a)?t.onClick?.(`month`,n.current,X(s),!1):t.resetDisplayedMonth())}),Z(`contextmenu`,h,function(...e){d.context?.apply(this,e)}),zr(`pointerenter`,h,function(...e){d.hover?.apply(this,e)}),Z(`click`,v,function(...e){f.click?.apply(this,e)}),Z(`keydown`,v,e=>{(e.key===`Enter`||e.key===` `)&&X(o)&&t.onClick?.(`year`,n.current,X(c),!1)}),Z(`contextmenu`,v,function(...e){f.context?.apply(this,e)}),zr(`pointerenter`,v,function(...e){f.hover?.apply(this,e)}),$(e,p),Tt()}Br([`click`,`keydown`,`contextmenu`]);var Ri=Q(`<div class="nav svelte-17quior"><!> <div class="right-nav svelte-17quior"><!> <button type="button">&#x25CF;</button> <!></div></div>`),zi={hash:`svelte-17quior`,code:`.nav.svelte-17quior {align-items:baseline;display:flex;margin:0.6em 0 1em;padding:0 8px;width:100%;}.right-nav.svelte-17quior {align-items:center;display:flex;justify-content:center;margin-left:auto;}.reset-button.svelte-17quior {align-items:center;appearance:none;background:none;border:none;color:var(--color-arrow);cursor:default;display:flex;font-size:8px;opacity:0.4;padding:0.5em;}.reset-button.active.svelte-17quior {cursor:pointer;opacity:1;}`};function Bi(e,t){wt(t,!0),di(e,zi);let n=St(u);function r(){n.current=n.current.clone().add(1,`month`)}function i(){n.current=n.current.clone().subtract(1,`month`)}function a(){n.current=t.today.clone()}let o=F(()=>n.current.isSame(t.today,`month`));var s=Ri(),c=z(s);Li(c,Ci({get fileMap(){return t.fileMap},resetDisplayedMonth:a},()=>t.eventHandlers));var l=Ln(c,2),d=z(l);Pi(d,{direction:`left`,onClick:i,tooltip:`Previous Month`});var f=Ln(d,2);let p;Pi(Ln(f,2),{direction:`right`,onClick:r,tooltip:`Next Month`}),O(l),O(s),er(()=>{vi(f,`aria-label`,X(o)?`Current month`:`Reset to current month`),p=mi(f,1,`reset-button svelte-17quior`,null,p,{active:!X(o)})}),Z(`click`,f,a),$(e,s),Tt()}Br([`click`]);function Vi(){return window.moment.weekdaysShort(!0)}function Hi(e){return e.isoWeekday()===6||e.isoWeekday()===7}function Ui(e){return e[0].clone()}function Wi(e){let t=[],n,r=e.clone().date(1),i=r.weekday(),a=r.clone().subtract(i,`days`);for(let e=0;e<42;e++)e%7==0&&(n={days:[],weekNum:a.week()},t.push(n)),n.days.push(a),a=a.clone().add(1,`days`);return t}var Gi=Q(`<td class="svelte-uujyui"><div role="button" tabindex="0"> </div></td>`),Ki={hash:`svelte-uujyui`,code:`td.svelte-uujyui {border-right:1px solid var(--background-modifier-border);}.week-num.svelte-uujyui {background-color:var(--color-background-weeknum);border-radius:4px;color:var(--color-text-weeknum);cursor:pointer;font-size:0.65em;height:100%;padding:4px;text-align:center;transition:background-color 0.1s ease-in,
+      color 0.1s ease-in;vertical-align:baseline;}.week-num.svelte-uujyui:hover {background-color:var(--interactive-hover);}.week-num.active.svelte-uujyui:hover {background-color:var(--interactive-accent-hover);}.active.svelte-uujyui {color:var(--text-on-accent);background-color:var(--interactive-accent);}`};function qi(e,t){wt(t,!0),di(e,Ki);let n=wi(t,`activeFilePath`,3,null),r=F(()=>Ui(t.days)),i=F(()=>t.fileMap.get(Ei(`week`,X(r)))??null);function a(e){e.target&&t.onHover?.(`week`,X(r),X(i),e.target,Ti(e))}var o=Gi(),s=z(o);let c;var l=z(s,!0);O(s),O(o),er(()=>{c=mi(s,1,`week-num svelte-uujyui`,null,c,{active:X(i)!==null&&X(i).path===n()}),qr(l,t.weekNum)}),Z(`click`,s,function(...e){(t.onClick&&(e=>t.onClick(`week`,X(r),X(i),Ti(e))))?.apply(this,e)}),Z(`keydown`,s,function(...e){(t.onClick&&(e=>(e.key===`Enter`||e.key===` `)&&t.onClick(`week`,X(r),X(i),!1)))?.apply(this,e)}),Z(`contextmenu`,s,function(...e){(t.onContextMenu&&(e=>t.onContextMenu(`week`,X(r),X(i),e)))?.apply(this,e)}),zr(`pointerenter`,s,a),$(e,o),Tt()}Br([`click`,`keydown`,`contextmenu`]);var Ji=Q(`<col/>`),Yi=Q(`<col/>`),Xi=Q(`<th class="svelte-5j20fs">W</th>`),Zi=Q(`<th class="svelte-5j20fs"> </th>`),Qi=Q(`<tr><!><!></tr>`),$i=Q(`<div id="calendar-container" class="container svelte-5j20fs"><!> <table class="calendar svelte-5j20fs"><colgroup><!><!></colgroup><thead><tr><!><!></tr></thead><tbody></tbody></table></div>`),ea={hash:`svelte-5j20fs`,code:`.container.svelte-5j20fs {--color-background-heading: transparent;--color-background-day: transparent;--color-background-weeknum: transparent;--color-background-weekend: transparent;--color-arrow: var(--text-muted);--color-button: var(--text-muted);--color-text-title: var(--text-normal);--color-text-heading: var(--text-muted);--color-text-day: var(--text-normal);--color-text-today: var(--interactive-accent);--color-text-weeknum: var(--text-muted);}.container.svelte-5j20fs {padding:0 8px;}.weekend.svelte-5j20fs {background-color:var(--color-background-weekend);}.calendar.svelte-5j20fs {border-collapse:collapse;width:100%;}th.svelte-5j20fs {background-color:var(--color-background-heading);color:var(--color-text-heading);font-size:0.6em;letter-spacing:1px;padding:4px;text-align:center;text-transform:uppercase;}`};function ta(e,t){wt(t,!0),di(e,ea);let n=I(null),r=I(window.moment()),i=new ji;Ct(u,i);let a=I(Wi(window.moment())),o=I(!1),s=I(new Map);Jn(()=>{L(a,Wi(i.current))}),Jn(()=>{t.fileStore.version,L(o,t.fileStore.isGranularityEnabled(`week`),!0),L(s,Di(X(a),(e,n)=>t.fileStore.getFile(e,n),t.fileStore.getEnabledGranularities()))});let c=F(()=>({onHover:t.onHover,onClick:t.onClick,onContextMenu:t.onContextMenu})),l=Vi();function d(){let e=window.moment();e.isSame(X(r),`day`)||L(r,e)}function f(e){L(n,e,!0)}var p={tick:d,setActiveFilePath:f},m=$i(),h=z(m);Bi(h,{get fileMap(){return X(s)},get today(){return X(r)},get eventHandlers(){return X(c)}});var g=Ln(h,2),_=z(g),v=z(_),y=e=>{$(e,Ji())};ei(v,e=>{X(o)&&e(y)}),ai(Ln(v),17,()=>X(a)[1].days,ti,(e,t)=>{var n=Yi();let r;er(e=>r=mi(n,1,`svelte-5j20fs`,null,r,e),[()=>({weekend:Hi(X(t))})]),$(e,n)}),O(_);var b=Ln(_),ee=z(b),x=z(ee),te=e=>{$(e,Xi())};ei(x,e=>{X(o)&&e(te)}),ai(Ln(x),17,()=>l,ti,(e,t)=>{var n=Zi(),r=z(n,!0);O(n),er(()=>qr(r,X(t))),$(e,n)}),O(ee),O(b);var ne=Ln(b);return ai(ne,21,()=>X(a),e=>Ei(`week`,e.days[0]),(e,t)=>{var i=Qi(),a=z(i),l=e=>{qi(e,Ci({get fileMap(){return X(s)},get activeFilePath(){return X(n)}},()=>X(t),()=>X(c)))};ei(a,e=>{X(o)&&e(l)}),ai(Ln(a),17,()=>X(t).days,e=>e.format(),(e,t)=>{Ai(e,Ci({get date(){return X(t)},get fileMap(){return X(s)},get today(){return X(r)},get activeFilePath(){return X(n)}},()=>X(c)))}),O(i),$(e,i)}),O(ne),O(g),O(m),$(e,m),Tt(p)}var na=class{#e=I(0);get version(){return X(this.#e)}set version(e){L(this.#e,e,!0)}plugin;constructor(e,t){this.plugin=t,t.app.workspace.onLayoutReady(()=>{let{vault:n,metadataCache:r,workspace:i}=t.app;e.registerEvent(n.on(`create`,this.bump,this)),e.registerEvent(n.on(`delete`,this.bumpUnconditionally,this)),e.registerEvent(n.on(`rename`,this.onRename,this)),e.registerEvent(r.on(`changed`,this.bump,this)),e.registerEvent(i.on(`periodic-notes:resolve`,this.bumpUnconditionally,this)),e.registerEvent(i.on(`periodic-notes:settings-updated`,this.bumpUnconditionally,this)),this.bump()})}bump(e){e&&!this.plugin.isPeriodic(e.path)||this.version++}bumpUnconditionally(){this.version++}onRename(e,t){(this.plugin.isPeriodic(e.path)||this.plugin.isPeriodic(t))&&this.version++}getFile(e,t){return this.plugin.getPeriodicNote(t,e)}isGranularityEnabled(e){return this.plugin.settings.granularities[e]?.enabled??e===`day`}getEnabledGranularities(){return[`week`,`month`,`year`].filter(e=>this.plugin.settings.granularities[e]?.enabled)}},ra=class extends e.ItemView{calendar;plugin;constructor(e,t){super(e),this.plugin=t,this.registerEvent(this.app.workspace.on(`file-open`,this.onFileOpen.bind(this)))}getViewType(){return l}getDisplayText(){return`Calendar`}getIcon(){return`calendar-day`}async onClose(){this.calendar&&Qr(this.calendar)}async onOpen(){let e=new na(this,this.plugin),t=Jr(ta,{target:this.contentEl,props:{fileStore:e,onHover:this.onHover.bind(this),onClick:this.onClick.bind(this),onContextMenu:this.onContextMenu.bind(this)}});if(!(`tick`in t&&`setActiveFilePath`in t))throw Error(`Calendar component missing expected exports`);this.calendar=t}onHover(e,t,n,r,i){if(!i)return;let a=t.format(e===`day`?`YYYY-MM-DD`:t.localeData().longDateFormat(`L`));this.app.workspace.trigger(`link-hover`,this,r,a,n?.path??``)}onClick(e,t,n,r){this.plugin.openPeriodicNote(e,t,{inNewSplit:r})}onContextMenu(t,n,r,i){if(!r)return;let a=new e.Menu;a.addItem(e=>e.setTitle(`Delete`).setIcon(`trash`).onClick(()=>{this.app.vault.trash(r,!0)})),this.app.workspace.trigger(`file-menu`,a,r,`calendar-context-menu`,null),a.showAtPosition({x:i.pageX,y:i.pageY})}onFileOpen(e){if(this.app.workspace.layoutReady&&this.calendar){let e=this.app.workspace.getActiveFile()?.path??null;this.calendar.setActiveFilePath(e),this.calendar.tick()}}},ia={day:{periodicity:`daily`,relativeUnit:`today`,labelOpenPresent:`Open today's daily note`},week:{periodicity:`weekly`,relativeUnit:`this week`,labelOpenPresent:`Open this week's note`},month:{periodicity:`monthly`,relativeUnit:`this month`,labelOpenPresent:`Open this month's note`},year:{periodicity:`yearly`,relativeUnit:`this year`,labelOpenPresent:`Open this year's note`}};async function aa(t,n,r){let i=t.workspace.getActiveFile();if(!i)return;let a=n.findInCache(i.path);if(!a)return;let o=n.findAdjacent(i.path,r);if(o){let n=t.vault.getAbstractFileByPath(o.filePath);n&&n instanceof e.TFile&&await t.workspace.getLeaf().openFile(n,{active:!0})}else{let t=r===`forwards`?`after`:`before`;new e.Notice(`There's no ${ia[a.granularity].periodicity} note ${t} this`)}}async function oa(e,t,n){let r=e.workspace.getActiveFile();if(!r)return;let i=t.findInCache(r.path);if(!i)return;let a=n===`forwards`?1:-1,o=i.date.clone().add(a,i.granularity);t.openPeriodicNote(i.granularity,o)}function sa(e,t,n){let r=ia[n];return[{id:`open-${r.periodicity}-note`,name:r.labelOpenPresent,checkCallback:e=>{if(!t.settings.granularities[n].enabled)return!1;if(e)return!0;t.openPeriodicNote(n,window.moment())}},{id:`next-${r.periodicity}-note`,name:`Jump forwards to closest ${r.periodicity} note`,checkCallback:r=>{if(!t.settings.granularities[n].enabled)return!1;let i=e.workspace.getActiveFile();if(r)return i?t.isPeriodic(i.path,n):!1;aa(e,t,`forwards`)}},{id:`prev-${r.periodicity}-note`,name:`Jump backwards to closest ${r.periodicity} note`,checkCallback:r=>{if(!t.settings.granularities[n].enabled)return!1;let i=e.workspace.getActiveFile();if(r)return i?t.isPeriodic(i.path,n):!1;aa(e,t,`backwards`)}},{id:`open-next-${r.periodicity}-note`,name:`Open next ${r.periodicity} note`,checkCallback:r=>{if(!t.settings.granularities[n].enabled)return!1;let i=e.workspace.getActiveFile();if(r)return i?t.isPeriodic(i.path,n):!1;oa(e,t,`forwards`)}},{id:`open-prev-${r.periodicity}-note`,name:`Open previous ${r.periodicity} note`,checkCallback:r=>{if(!t.settings.granularities[n].enabled)return!1;let i=e.workspace.getActiveFile();if(r)return i?t.isPeriodic(i.path,n):!1;oa(e,t,`backwards`)}}]}function ca(t,n){let i=new e.Menu,a=r.filter(e=>t.settings.granularities[e].enabled);for(let e of a){let n=ia[e];i.addItem(r=>r.setTitle(n.labelOpenPresent).setIcon(`calendar-${e}`).onClick(()=>t.openPeriodicNote(e,window.moment())))}i.showAtPosition(n)}var la=`
 <g>
 <path d="M24.78 3C22.646 3 20.9 4.746 20.9 6.88V10.76H9.26C7.223 10.76 5.38 12.312 5.38 14.543V92.628C5.38 93.695 6.059 94.859 6.835 95.344C7.611 95.926 8.387 96.12 9.26 96.12H90.74C91.613 96.12 92.389 95.926 93.165 95.344C93.941 94.762 94.62 93.695 94.62 92.628V14.543C94.62 12.506 92.971 10.76 90.934 10.76H79.1V6.88C79.1 4.746 77.354 3 75.22 3H71.34C69.206 3 67.46 4.746 67.46 6.88V10.76H32.54V6.88C32.54 4.746 30.794 3 28.66 3H24.78ZM24.78 6.88H28.66V18.52H24.78V6.88ZM71.34 6.88H75.22V18.52H71.34V6.88ZM9.26 14.64H20.9V18.52C20.9 20.654 22.646 22.4 24.78 22.4H28.66C30.794 22.4 32.54 20.654 32.54 18.52V14.64H67.46V18.52C67.46 20.654 69.206 22.4 71.34 22.4H75.22C77.354 22.4 79.1 20.654 79.1 18.52V14.64H90.74V28.22H9.26V14.64ZM9.26 32.1H90.74V92.24H9.26V32.1Z" fill="currentColor" stroke-width="1" stroke="currentColor"/>
 <path d="M55.2539 79.0024H49.3613V55.3319C49.3613 52.5068 49.4282 50.2668 49.5619 48.6119C49.1775 49.0131 48.701 49.4561 48.1327 49.9408C47.581 50.4256 45.7088 51.9635 42.516 54.5546L39.5571 50.8185L50.3393 42.3432H55.2539V79.0024Z" fill="currentColor"/>
 </g>
-`,fa=`
+`,ua=`
 <g>
 <path d="M24.78 3C22.646 3 20.9 4.746 20.9 6.88V10.76H9.26C7.223 10.76 5.38 12.312 5.38 14.543V92.628C5.38 93.695 6.059 94.859 6.835 95.344C7.611 95.926 8.387 96.12 9.26 96.12H90.74C91.613 96.12 92.389 95.926 93.165 95.344C93.941 94.762 94.62 93.695 94.62 92.628V14.543C94.62 12.506 92.971 10.76 90.934 10.76H79.1V6.88C79.1 4.746 77.354 3 75.22 3H71.34C69.206 3 67.46 4.746 67.46 6.88V10.76H32.54V6.88C32.54 4.746 30.794 3 28.66 3H24.78ZM24.78 6.88H28.66V18.52H24.78V6.88ZM71.34 6.88H75.22V18.52H71.34V6.88ZM9.26 14.64H20.9V18.52C20.9 20.654 22.646 22.4 24.78 22.4H28.66C30.794 22.4 32.54 20.654 32.54 18.52V14.64H67.46V18.52C67.46 20.654 69.206 22.4 71.34 22.4H75.22C77.354 22.4 79.1 20.654 79.1 18.52V14.64H90.74V28.22H9.26V14.64ZM9.26 32.1H90.74V92.24H9.26V32.1Z" fill="currentColor" stroke-width="1" stroke="currentColor"/>
 <path d="M42.8799 78.3604L56.5679 48.6873H38.5698V43.7852H62.512V47.669L48.895 78.3604H42.8799Z" fill="currentColor"/>
 </g>
-`,pa=`
+`,da=`
 <g>
 <path d="M24.78 3C22.646 3 20.9 4.746 20.9 6.88V10.76H9.26C7.223 10.76 5.38 12.312 5.38 14.543V92.628C5.38 93.695 6.059 94.859 6.835 95.344C7.611 95.926 8.387 96.12 9.26 96.12H90.74C91.613 96.12 92.389 95.926 93.165 95.344C93.941 94.762 94.62 93.695 94.62 92.628V14.543C94.62 12.506 92.971 10.76 90.934 10.76H79.1V6.88C79.1 4.746 77.354 3 75.22 3H71.34C69.206 3 67.46 4.746 67.46 6.88V10.76H32.54V6.88C32.54 4.746 30.794 3 28.66 3H24.78ZM24.78 6.88H28.66V18.52H24.78V6.88ZM71.34 6.88H75.22V18.52H71.34V6.88ZM9.26 14.64H20.9V18.52C20.9 20.654 22.646 22.4 24.78 22.4H28.66C30.794 22.4 32.54 20.654 32.54 18.52V14.64H67.46V18.52C67.46 20.654 69.206 22.4 71.34 22.4H75.22C77.354 22.4 79.1 20.654 79.1 18.52V14.64H90.74V28.22H9.26V14.64ZM9.26 32.1H90.74V92.24H9.26V32.1Z" fill="currentColor" stroke-width="1" stroke="currentColor"/>
 <path d="M51.3075 52.8546C51.3075 54.9201 50.7057 56.6437 49.5022 58.0256C48.2986 59.3926 46.6046 60.3139 44.4204 60.7894V60.9677C47.0356 61.2946 48.9969 62.1118 50.3045 63.4194C51.6121 64.7122 52.2659 66.4358 52.2659 68.5904C52.2659 71.7257 51.1589 74.1477 48.9449 75.8565C46.7309 77.5504 43.5808 78.3974 39.4946 78.3974C35.8838 78.3974 32.8377 77.8105 30.3562 76.6366V71.9783C31.7381 72.6618 33.2018 73.1893 34.7471 73.5608C36.2924 73.9322 37.7784 74.118 39.2048 74.118C41.7309 74.118 43.618 73.6499 44.8661 72.7138C46.1143 71.7777 46.7384 70.3289 46.7384 68.3675C46.7384 66.629 46.0474 65.3511 44.6655 64.5339C43.2836 63.7166 41.1142 63.308 38.1573 63.308H35.3266V59.0509H38.2018C43.4025 59.0509 46.0028 57.2529 46.0028 53.657C46.0028 52.2603 45.5496 51.183 44.6432 50.4252C43.7368 49.6674 42.3995 49.2885 40.6313 49.2885C39.398 49.2885 38.2093 49.4668 37.0651 49.8234C35.921 50.1652 34.5688 50.8412 33.0086 51.8517L30.4454 48.1963C33.4321 45.9972 36.9017 44.8976 40.8542 44.8976C44.138 44.8976 46.7012 45.6034 48.5437 47.015C50.3863 48.4266 51.3075 50.3732 51.3075 52.8546Z" fill="currentColor"/>
 <path d="M69.6199 77.9516H64.382V56.9112C64.382 54.4 64.4415 52.4089 64.5603 50.9378C64.2186 51.2944 63.7951 51.6882 63.2899 52.1191C62.7995 52.55 61.1353 53.9171 58.2972 56.2202L55.6672 52.8992L65.2513 45.3657H69.6199V77.9516Z" fill="currentColor"/>
 </g>
-`,ma=`
+`,fa=`
 <g>
 <path d="M24.768 3C22.634 3 20.888 4.746 20.888 6.88V10.76H9.24804C7.21104 10.76 5.36804 12.312 5.36804 14.543V92.628C5.36804 93.695 6.04704 94.859 6.82304 95.344C7.59904 95.926 8.37504 96.12 9.24804 96.12H90.728C91.601 96.12 92.377 95.926 93.153 95.344C93.929 94.762 94.608 93.695 94.608 92.628V14.543C94.608 12.506 92.959 10.76 90.922 10.76H79.088V6.88C79.088 4.746 77.342 3 75.208 3H71.328C69.194 3 67.448 4.746 67.448 6.88V10.76H32.528V6.88C32.528 4.746 30.782 3 28.648 3H24.768ZM24.768 6.88H28.648V18.52H24.768V6.88ZM71.328 6.88H75.208V18.52H71.328V6.88ZM9.24804 14.64H20.888V18.52C20.888 20.654 22.634 22.4 24.768 22.4H28.648C30.782 22.4 32.528 20.654 32.528 18.52V14.64H67.448V18.52C67.448 20.654 69.194 22.4 71.328 22.4H75.208C77.342 22.4 79.088 20.654 79.088 18.52V14.64H90.728V28.22H9.24804V14.64ZM9.24804 32.1H90.728V92.24H9.24804V32.1Z" fill="currentColor" stroke="currentColor" stroke-width="0.17"/>
 <path d="M49.2303 60.2321L56.9421 45.3656H62.7371L51.8826 65.3139V77.9515H46.5333V65.4922L35.7234 45.3656H41.5184L49.2303 60.2321Z" fill="currentColor"/>
 </g>
-`,ha=class extends e.AbstractInputSuggest{onSelectCallback;constructor(e,t,n){super(e,t),this.onSelectCallback=n}getSuggestions(e){let t=e.toLowerCase();return this.app.vault.getMarkdownFiles().filter(e=>e.path.toLowerCase().contains(t))}renderSuggestion(e,t){t.setText(e.path)}selectSuggestion(e){this.setValue(e.path),this.onSelectCallback?.(e.path),this.close()}},ga=class extends e.AbstractInputSuggest{onSelectCallback;constructor(e,t,n){super(e,t),this.onSelectCallback=n}getSuggestions(e){let t=e.toLowerCase();return this.app.vault.getAllFolders().filter(e=>e.path.toLowerCase().contains(t))}renderSuggestion(e,t){t.setText(e.path)}selectSuggestion(e){this.setValue(e.path),this.onSelectCallback?.(e.path),this.close()}};function _a(e,t){return t?e.metadataCache.getFirstLinkpathDest(t,``)?``:`Template file not found`:``}function va(t,n){return!n||n===`/`||t.vault.getAbstractFileByPath((0,e.normalizePath)(n))?``:`Folder not found in vault`}var ya={day:`Daily Notes`,week:`Weekly Notes`,month:`Monthly Notes`,year:`Yearly Notes`},ba=class extends e.PluginSettingTab{debouncedSave=(0,e.debounce)(()=>this.plugin.saveSettings(),500,!0);constructor(e,t){super(e,t),this.app=e,this.plugin=t}display(){let{containerEl:e}=this;e.empty();for(let t of r)this.addGranularitySection(e,t)}addGranularitySection(t,n){let r=this.plugin.settings.granularities[n];t.createEl(`h3`,{text:ya[n]}),new e.Setting(t).setName(`Enabled`).addToggle(e=>e.setValue(r.enabled).onChange(async e=>{this.plugin.settings.granularities[n].enabled=e,await this.plugin.saveSettings()}));let i=new e.Setting(t).setName(`Format`).setDesc(`Moment.js date format string`).addText(e=>{e.setPlaceholder(a[n]).setValue(r.format).onChange(async e=>{let t=v(e,n);i.descEl.setText(t||`Moment.js date format string`),i.descEl.toggleClass(`has-error`,!!t),this.plugin.settings.granularities[n].format=e,this.debouncedSave()})}),o=new e.Setting(t).setName(`Folder`).addText(e=>{e.setValue(r.folder).onChange(async e=>{let t=va(this.app,e);o.descEl.setText(t||``),o.descEl.toggleClass(`has-error`,!!t),this.plugin.settings.granularities[n].folder=e,this.debouncedSave()}),new ga(this.app,e.inputEl)}),s=new e.Setting(t).setName(`Template`).addText(e=>{e.setValue(r.templatePath??``).onChange(async e=>{let t=_a(this.app,e);s.descEl.setText(t||``),s.descEl.toggleClass(`has-error`,!!t),this.plugin.settings.granularities[n].templatePath=e||void 0,this.debouncedSave()}),new ha(this.app,e.inputEl)})}},xa={en:`en-gb`,zh:`zh-cn`,"zh-TW":`zh-tw`,ru:`ru`,ko:`ko`,it:`it`,id:`id`,ro:`ro`,"pt-BR":`pt-br`,cz:`cs`,da:`da`,de:`de`,es:`es`,fr:`fr`,no:`nn`,pl:`pl`,pt:`pt`,tr:`tr`,hi:`hi`,nl:`nl`,ar:`ar`,ja:`ja`};function Sa(){let e=localStorage.getItem(`language`)||`en`,t=navigator.language?.toLowerCase(),n=xa[e]??e;t?.startsWith(e)&&(n=t);let r=window.moment.locale(n);console.debug(`[Periodic Notes] Configured locale: requested ${n}, got ${r}`)}var Ca=class extends e.Plugin{settings;ribbonEl;cache;async onload(){(0,e.addIcon)(`calendar-day`,da),(0,e.addIcon)(`calendar-week`,fa),(0,e.addIcon)(`calendar-month`,pa),(0,e.addIcon)(`calendar-year`,ma),await this.loadSettings(),Sa(),this.ribbonEl=null,this.cache=new me(this.app,this),this.openPeriodicNote=this.openPeriodicNote.bind(this),this.addSettingTab(new ba(this.app,this)),this.configureRibbonIcons(),this.configureCommands(),this.registerView(l,e=>new aa(e,this)),this.addCommand({id:`show-calendar`,name:`Show calendar`,checkCallback:e=>{if(e)return this.app.workspace.getLeavesOfType(l).length===0;this.app.workspace.getRightLeaf(!1)?.setViewState({type:l})}})}configureRibbonIcons(){this.ribbonEl?.detach();let e=r.find(e=>this.settings.granularities[e].enabled);if(e){let t=oa[e];this.ribbonEl=this.addRibbonIcon(`calendar-${e}`,t.labelOpenPresent,t=>{t.type!==`auxclick`&&this.openPeriodicNote(e,window.moment(),{inNewSplit:Di(t)})}),this.ribbonEl.addEventListener(`contextmenu`,e=>{e.preventDefault(),ua(this,{x:e.pageX,y:e.pageY})})}}configureCommands(){for(let e of r)la(this.app,this,e).forEach(this.addCommand.bind(this))}async loadSettings(){let e=await this.loadData(),t=structuredClone(s);if(e?.granularities)for(let n of r)e.granularities[n]&&(t.granularities[n]={...t.granularities[n],...e.granularities[n]});this.settings=t}async saveSettings(){await this.saveData(this.settings),this.configureRibbonIcons(),this.app.workspace.trigger(`periodic-notes:settings-updated`)}async createPeriodicNote(e,t){let n=p(this.settings,e),r=d(this.settings,e),i=t.format(r),a=se(i,e,t,r,await ce(this.app,n.templatePath,e)),o=await ue(this.app,i,n);return this.app.vault.create(o,a)}getPeriodicNote(e,t){return this.cache.getPeriodicNote(e,t)}isPeriodic(e,t){return this.cache.isPeriodic(e,t)}findAdjacent(e,t){return this.cache.findAdjacent(e,t)}findInCache(e){return this.cache.find(e)}async openPeriodicNote(e,t,n){let{inNewSplit:r=!1}=n??{},{workspace:i}=this.app,a=this.cache.getPeriodicNote(e,t);a||=await this.createPeriodicNote(e,t),await(r?i.getLeaf(`split`):i.getLeaf()).openFile(a,{active:!0})}};module.exports=Ca;
+`,pa=class extends e.AbstractInputSuggest{onSelectCallback;constructor(e,t,n){super(e,t),this.onSelectCallback=n}getSuggestions(e){let t=e.toLowerCase();return this.app.vault.getMarkdownFiles().filter(e=>e.path.toLowerCase().contains(t))}renderSuggestion(e,t){t.setText(e.path)}selectSuggestion(e){this.setValue(e.path),this.onSelectCallback?.(e.path),this.close()}},ma=class extends e.AbstractInputSuggest{onSelectCallback;constructor(e,t,n){super(e,t),this.onSelectCallback=n}getSuggestions(e){let t=e.toLowerCase();return this.app.vault.getAllFolders().filter(e=>e.path.toLowerCase().contains(t))}renderSuggestion(e,t){t.setText(e.path)}selectSuggestion(e){this.setValue(e.path),this.onSelectCallback?.(e.path),this.close()}};function ha(e,t){return t?e.metadataCache.getFirstLinkpathDest(t,``)?``:`Template file not found`:``}function ga(t,n){return!n||n===`/`||t.vault.getAbstractFileByPath((0,e.normalizePath)(n))?``:`Folder not found in vault`}var _a={day:`Daily Notes`,week:`Weekly Notes`,month:`Monthly Notes`,year:`Yearly Notes`},va=class extends e.PluginSettingTab{debouncedSave=(0,e.debounce)(()=>this.plugin.saveSettings(),500,!0);constructor(e,t){super(e,t),this.app=e,this.plugin=t}display(){let{containerEl:e}=this;e.empty();for(let t of r)this.addGranularitySection(e,t)}addGranularitySection(t,n){let r=this.plugin.settings.granularities[n];t.createEl(`h3`,{text:_a[n]}),new e.Setting(t).setName(`Enabled`).addToggle(e=>e.setValue(r.enabled).onChange(async e=>{this.plugin.settings.granularities[n].enabled=e,await this.plugin.saveSettings()}));let i=new e.Setting(t).setName(`Format`).setDesc(`Moment.js date format string`).addText(e=>{e.setPlaceholder(a[n]).setValue(r.format).onChange(async e=>{let t=v(e,n);i.descEl.setText(t||`Moment.js date format string`),i.descEl.toggleClass(`has-error`,!!t),this.plugin.settings.granularities[n].format=e,this.debouncedSave()})}),o=new e.Setting(t).setName(`Folder`).addText(e=>{e.setValue(r.folder).onChange(async e=>{let t=ga(this.app,e);o.descEl.setText(t||``),o.descEl.toggleClass(`has-error`,!!t),this.plugin.settings.granularities[n].folder=e,this.debouncedSave()}),new ma(this.app,e.inputEl)}),s=new e.Setting(t).setName(`Template`).addText(e=>{e.setValue(r.templatePath??``).onChange(async e=>{let t=ha(this.app,e);s.descEl.setText(t||``),s.descEl.toggleClass(`has-error`,!!t),this.plugin.settings.granularities[n].templatePath=e||void 0,this.debouncedSave()}),new pa(this.app,e.inputEl)})}},ya={en:`en-gb`,zh:`zh-cn`,"zh-TW":`zh-tw`,ru:`ru`,ko:`ko`,it:`it`,id:`id`,ro:`ro`,"pt-BR":`pt-br`,cz:`cs`,da:`da`,de:`de`,es:`es`,fr:`fr`,no:`nn`,pl:`pl`,pt:`pt`,tr:`tr`,hi:`hi`,nl:`nl`,ar:`ar`,ja:`ja`};function ba(){let e=localStorage.getItem(`language`)||`en`,t=navigator.language?.toLowerCase(),n=ya[e]??e;t?.startsWith(e)&&(n=t);let r=window.moment.locale(n);console.debug(`[Periodic Notes] Configured locale: requested ${n}, got ${r}`)}var xa=class extends e.Plugin{settings;ribbonEl;cache;async onload(){(0,e.addIcon)(`calendar-day`,la),(0,e.addIcon)(`calendar-week`,ua),(0,e.addIcon)(`calendar-month`,da),(0,e.addIcon)(`calendar-year`,fa),await this.loadSettings(),ba(),this.ribbonEl=null,this.cache=new me(this.app,this),this.openPeriodicNote=this.openPeriodicNote.bind(this),this.addSettingTab(new va(this.app,this)),this.configureRibbonIcons(),this.configureCommands(),this.registerView(l,e=>new ra(e,this)),this.addCommand({id:`show-calendar`,name:`Show calendar`,checkCallback:e=>{if(e)return this.app.workspace.getLeavesOfType(l).length===0;this.app.workspace.getRightLeaf(!1)?.setViewState({type:l})}})}configureRibbonIcons(){this.ribbonEl?.detach();let e=r.find(e=>this.settings.granularities[e].enabled);if(e){let t=ia[e];this.ribbonEl=this.addRibbonIcon(`calendar-${e}`,t.labelOpenPresent,t=>{t.type!==`auxclick`&&this.openPeriodicNote(e,window.moment(),{inNewSplit:Ti(t)})}),this.ribbonEl.addEventListener(`contextmenu`,e=>{e.preventDefault(),ca(this,{x:e.pageX,y:e.pageY})})}}configureCommands(){for(let e of r)sa(this.app,this,e).forEach(this.addCommand.bind(this))}async loadSettings(){let e=await this.loadData(),t=structuredClone(s);if(e?.granularities)for(let n of r)e.granularities[n]&&(t.granularities[n]={...t.granularities[n],...e.granularities[n]});this.settings=t}async saveSettings(){await this.saveData(this.settings),this.configureRibbonIcons(),this.app.workspace.trigger(`periodic-notes:settings-updated`)}async createPeriodicNote(e,t){let n=p(this.settings,e),r=d(this.settings,e),i=t.format(r),a=se(i,e,t,r,await ce(this.app,n.templatePath,e)),o=await ue(this.app,i,n);return this.app.vault.create(o,a)}getPeriodicNote(e,t){return this.cache.getPeriodicNote(e,t)}isPeriodic(e,t){return this.cache.isPeriodic(e,t)}findAdjacent(e,t){return this.cache.findAdjacent(e,t)}findInCache(e){return this.cache.find(e)}async openPeriodicNote(e,t,n){let{inNewSplit:r=!1}=n??{},{workspace:i}=this.app,a=this.cache.getPeriodicNote(e,t);a||=await this.createPeriodicNote(e,t),await(r?i.getLeaf(`split`):i.getLeaf()).openFile(a,{active:!0})}};module.exports=xa;
